% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad-data.R
\docType{data}
\name{ad_ge_micro_zscore}
\alias{ad_ge_micro_zscore}
\title{Atopic dermatitis data}
\format{An object of class \code{data.frame}}
\usage{
data(ad_ge_micro_zscore)
}
\description{
Standardized (z-score)
}
\keyword{datasets}
