% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ad-data.R
\docType{data}
\name{ad_studies}
\alias{ad_studies}
\title{Atopic dermatitis studies}
\format{An object of class \code{factor}}
\usage{
data(ad_studies)
}
\description{
Study ID from annotations
}
\keyword{datasets}
