% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BankingCrisesDistances-data.R
\name{BankingCrisesDistances}
\alias{BankingCrisesDistances}
\title{Banking Crises Distances}
\format{A 69 x 70 matrix.}
\source{
data(bankingCrises) in library(Ecdat)
}
\description{
Matrix of Jaccard distances between 70 countries (Hungary and Greece were combined to be the same observation) based on their binary time series of having had a banking crises in a year from 1800 to 2010 or not. See data(bankingCrises) in pakcage Ecdat for more info. The last column is Reinhart & Rogoffs classification as a low (3), middle- (2) or high-income country (1).
}

