% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{bootstrap}
\alias{bootstrap}
\title{Subsampling permutations of clustering dataset}
\usage{
bootstrap(
  dat_list,
  nruns = 100,
  stratified_cv = FALSE,
  anti_stratified = FALSE,
  cv_stratification_var = NULL,
  extra_fold = TRUE
)
}
\value{
list of data.frames with added columns "fold", "run" and "cv_index" as well as 
        duplicated rows of the original data corresponding to different folds.
}
\description{
Creates subsamples via bootstrapping.
}
