% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brca-data.R
\docType{data}
\name{brca_multi_omics}
\alias{brca_multi_omics}
\title{Breast cancer multi-omic data}
\format{
An object of class \code{list}
}
\usage{
data(brca_multi_omics)
}
\description{
A list of matrics corresponding to RNA-Seq TPM, average gene promoter 
methylation M-values and copy number variations from GISTIC 2 for genes 
involved in selected KEGG pathways. From the TCGA breast cancer dataset 
(https://doi.org/10.1038/nature11412).
}
\keyword{datasets}
