% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.cops.file.R
\name{clean.cops.file}
\alias{clean.cops.file}
\title{This function cleans the COPS casts, includings those
obtained in shallow waters}
\usage{
clean.cops.file(
  in.file = "",
  out.file = "cleaned.cops.csv",
  distance.above.bottom = 0.15
)
}
\arguments{
\item{in.file}{is the file name of the raw COPS cast to process}

\item{out.file}{is the file name of the cleaned COPS file.
If NA, the input file will be overwrite. The default is "cleaned.cops.csv"}

\item{distance.above.bottom}{is the distance above the bottom at which the user want to cut the data.
This is for shallow water casts for which the last COPS measurement are affected by the instrument shadow.
If NA, we assume the profile in deep open water and all data are kept.
The default is 0.15, so the last 15 cm of the profile is removed.}
}
\description{
The function prompts the user to click on the begining and the end of the cast.
In shallow waters, the user should click at the depth where the cops
it the bottom. The measurements taken between that depth and the
depth specified by distance.above.bottom will be remove.
}
\author{
Simon Belanger
}
