% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{clinical_associations}
\alias{clinical_associations}
\title{Clinical association for clustering results}
\usage{
clinical_associations(clust, clinical_data)
}
\arguments{
\item{clust}{data.frame with columns id and cluster.}

\item{clinical_data}{data.frame with clinical variables, rownames should match with id in \code{clust}.}
}
\value{
A data.frame of NMI, ARI and Chi-squared test p-values for categorical variables 
        as well as analysis of variance and kruskal-wallis test p-values for 
        continuous variables.
}
\description{
Clinical association for clustering results
}
