% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{clinical_evaluation}
\alias{clinical_evaluation}
\title{Gene module correlation score for multiple clustering results}
\usage{
clinical_evaluation(
  clusters,
  clinical_data,
  by = c("run", "fold", "datname", "drname", "k", "m"),
  ...
)
}
\arguments{
\item{clusters}{A data.table or data.frame with clustering information.}

\item{clinical_data}{A data.frame with clinical variables (categoric or numeric).}

\item{by}{Column names that identify a single clustering result.}

\item{...}{Extra arguments are ignored.}
}
\value{
\code{data.table} containing clinical associations
}
\description{
Runs \code{\link{gene_module_score}} in parallel.
}
\details{
Assumes that a parallel backend has been registered for \code{foreach}.
}
