% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{cluster_associations}
\alias{cluster_associations}
\title{Association analysis for clustering results}
\usage{
cluster_associations(clusters, association_data)
}
\arguments{
\item{clusters}{data.frame with columns id and cluster.}

\item{association_data}{data.frame with association variables, rownames should match with id in \code{clusters}.}
}
\value{
A data.frame of NMI, ARI and Chi-squared test p-values for categorical variables 
        as well as analysis of variance and kruskal-wallis test p-values for 
        continuous variables.
}
\description{
Association analysis for clustering results
}
