% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering_analysis}
\alias{clustering_analysis}
\title{Clustering, internal evaluation and batch effect estimation}
\usage{
clustering_analysis(
  dat,
  n_clusters = 2:5,
  cluster_methods = c("hierarchical", "diana", "kmeans"),
  clustering_dissimilarity = NULL,
  distance_metric = "euclidean",
  correlation_method = "spearman",
  hierarchical_linkage = "complete",
  kmeans_num_init = 100,
  kmeans_max_iters = 100,
  kmeans_tol = 1e-04,
  gmm_modelNames = NULL,
  gmm_shrinkage = 0.01,
  knn_neighbours = 30,
  knn_jaccard = TRUE,
  ...
)
}
\arguments{
\item{dat}{A data.frame with features on columns labeled as "dim[0-9]+", must also contain "id" column.}

\item{n_clusters}{A vector of integers, numbers of clusters to be generated.}

\item{cluster_methods}{A vector of clustering method names, see details for options.}

\item{clustering_dissimilarity}{A dissimilarity matrix used in some methods such as hierarchical clustering. Computed with \code{\link{clustering_dissimilarity_from_data}} if missing.}

\item{distance_metric}{Either "euclidean" or "correlation".}

\item{correlation_method}{Method for \code{\link[stats]{cor}}.}

\item{hierarchical_linkage}{See \code{\link[flashClust]{flashClust}}.}

\item{kmeans_num_init}{See \code{\link[ClusterR]{KMeans_rcpp}}.}

\item{kmeans_max_iters}{See \code{\link[ClusterR]{KMeans_rcpp}}.}

\item{kmeans_tol}{See \code{\link[ClusterR]{KMeans_rcpp}}.}

\item{gmm_modelNames}{Sepcifies model type for \code{\link[mclust]{Mclust}}}

\item{gmm_shrinkage}{Shrinkage parameter for \code{\link[mclust]{priorControl}}.}

\item{knn_neighbours}{number of nearest neighbours for community detection.}

\item{knn_jaccard}{computes shared neighbour weights with Jaccard ubdex if \code{TRUE}.}

\item{...}{extra arguments are ignored currently}
}
\value{
Returns a \code{list} containing clusters, metrics, and
        \code{\link[stats]{chisq.test}} p-values
        if batch_label was supplied
}
\description{
Single embedding or dataset evaluation
}
\details{
Supported clustering methods are:
\itemize{
\item "hierarchical" - agglomerative hierarchical clustering
\item "diana" - divisive hierarchical clustering analysis
\item "kmeans" - k-means++
\item "model" - Gaussian Mixture Models
\item "knn_communities" - Louvain community detection on shared k nearest neighbour graphs
\item "spectral" - spectral clustering
}
}
