% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering_dissimilarity}
\alias{clustering_dissimilarity}
\title{Dissimilarity matrix}
\usage{
clustering_dissimilarity(
  x,
  distance_metric = "euclidean",
  correlation_method = "spearman",
  ...
)
}
\arguments{
\item{x}{}

\item{distance_metric}{}

\item{correlation_method}{}

\item{...}{}
}
\value{

}
\description{
Dissimilarity matrix
}
