% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering_dissimilarity_from_data}
\alias{clustering_dissimilarity_from_data}
\title{Dissimilarity matrix}
\usage{
clustering_dissimilarity_from_data(
  x,
  distance_metric = "euclidean",
  correlation_method = "spearman",
  preprocess = FALSE,
  ...
)
}
\arguments{
\item{x}{input data \code{matrix}}

\item{distance_metric}{Either "euclidean" or "correlation"}

\item{correlation_method}{method for \code{\link[stats]{cor}}}

\item{preprocess}{set to \code{TRUE}if \code{x} is a \code{data.frame} that 
is processed within the COPS pipeline with data columns named "dim1", "dim2" and so on}

\item{...}{ignored}
}
\value{
\code{dist}-object
}
\description{
Dissimilarity matrix
}
