% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{clustering_evaluation}
\alias{clustering_evaluation}
\title{Clustering, internal evaluation and batch effect estimation}
\usage{
clustering_evaluation(
  dat,
  batch_label_names = NULL,
  subtype_label_names = NULL,
  n_clusters = 2:5,
  cluster_methods = c("hierarchical", "diana", "kmeans"),
  distance_metric = "euclidean",
  correlation_method = "spearman",
  hierarchical_linkage = "complete",
  gmm_modelNames = NULL,
  kmeans_num_init = 100,
  kmeans_max_iters = 100,
  kmeans_tol = 1e-04,
  gmm_shrinkage = 0.01,
  ...
)
}
\arguments{
\item{dat}{data matrix with samples on columns}

\item{batch_label_names}{character vector containing column names corresponding to batch labels}

\item{subtype_label_names}{}

\item{n_clusters}{vector of integers, numbers of clusters to be generated}

\item{cluster_methods}{vector of clustering method names, see details for options}

\item{distance_metric}{}

\item{correlation_method}{}

\item{hierarchical_linkage}{}

\item{gmm_modelNames}{}

\item{kmeans_num_init}{}

\item{kmeans_max_iters}{}

\item{kmeans_tol}{}

\item{...}{extra arguments are ignored currently}
}
\value{
Returns a \code{list} containing clusters, metrics, and
        \code{\link[stats]{chisq.test}} p-values
        if batch_label was supplied
}
\description{
Single embedding or dataset evaluation
}
\details{
Supported clustering methods are:
\itemize{
\item "hierarchical" -
\item "diana" -
\item "kmeans" -
\item "model" -
}
}
