% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clustering_metrics}
\alias{clustering_metrics}
\title{Internal evaluation of clustering results}
\usage{
clustering_metrics(
  x,
  dat = NULL,
  by = c("k", "m"),
  dissimilarity = NULL,
  cluster_size_table = TRUE,
  silhouette_min_cluster_size = 0,
  ...
)
}
\arguments{
\item{x}{A data.frame of clustering results.}

\item{dat}{A data.frame with data columns identified with "dim". Not required if dissimilarity is defined.}

\item{by}{vector of variable names to split by}

\item{dissimilarity}{a \code{dist} object to use in silhouette calculation, defaults to euclidean distance matrix if left \code{NULL}}

\item{cluster_size_table}{return cluster sizes if \code{TRUE}.}

\item{silhouette_min_cluster_size}{proportional cluster size threshold for merging into nearest neighbours' cluster for silhouette computation.}

\item{...}{extra arguments are passed to \code{\link{clustering_dissimilarity_from_data}}}
}
\value{
Returns a \code{list} containing metrics and cluster sizes
}
\description{
Single embedding or dataset evaluation
}
