% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{clusteval_cv}
\alias{clusteval_cv}
\title{Cross-validation of clustering evaluation}
\usage{
clusteval_cv(dat, batch_label = NULL, CVFOLDS = 2, RUNS = 10, ...)
}
\arguments{
\item{dat}{data matrix, samples on columns}

\item{batch_label}{vector or matrix where columns are categorical variables}

\item{CVFOLDS}{number of cross-validation folds}

\item{RUNS}{number of times to repeat cross-validation}

\item{...}{extra arguments are passed through to clustering_evaluation}
}
\value{
Returns a \code{list} of \code{data.frames} containing \code{\link{clustering_evaluation}} outputs for every
        combination of CV run, CV fold, clustering method and number of clusters
}
\description{
Performs cross-validation to estimate clustering stability and the spread of other metrics.
}
\details{
Stability of clustering should decrease with data size, which means that using
fewer folds for cross-validation is more difficult i.e. should be less stable.
}
