% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{coexpression_network_unweighted}
\alias{coexpression_network_unweighted}
\title{Unweighted gene co-expression network constructor}
\usage{
coexpression_network_unweighted(
  dat,
  correlation_method = "spearman",
  cor_threshold = 0.5
)
}
\arguments{
\item{dat}{gene expression data, samples on columns}

\item{correlation_method}{correlation method}

\item{cor_threshold}{numeric threshold used to define edges/links (see \code{\link[WGCNA]{pickHardThreshold}})}
}
\value{

}
\description{
Generates a gene co-expression network by thresholding gene expression correlations with \code{\link[WGCNA]{signumAdjacencyFunction}}.
}
