% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.Ksurf.linear.R
\name{compute.Ksurf.linear}
\alias{compute.Ksurf.linear}
\title{Extrapolate vertical profile of a radiometric
quantity (EdZ, EuZ, LuZ) to the air-sea interface using a linear
model applied to log-transformed data in an iterative scheme.}
\usage{
compute.Ksurf.linear(
  Depth,
  aop,
  r2.threshold = 0.8,
  instrument = "LuZ",
  delta.depth = 2.5
)
}
\arguments{
\item{Depth}{is a vector of depth corresponding to the input quantity}

\item{aop}{is the matrix of EdZ, EuZ or LuZ. The matrix should be clean
after filtration for high instrument tilt.}

\item{r2.threshold}{is the minimum R^2 require to accept the fit
(Default is 0.80)}

\item{delta.depth}{is the depth interval toletared between the
first and the last depth condidered for the surface.
The Default is 2.5 meters. So if the very first measurement is
at 0.5 m, the "surface" layer is considered as 0.5 to 3 meters depth.
But this is the maximum layer considered.
The minimum layer is fixed at 0.30 m.}
}
\value{
A list of list of 5 vectors of 19 values (one per wavelenght):
X.0m (is the quantity (EdZ, LuZ or EuZ) just below the sea surface),
Kx (is the diffuse attenuation coefficient of the surface layer considered),
Z.interval (is the maximum depth considered),
ix.Z.interval (is the index of the maximum depth considered),
r2 (is the R-squared of the regression),
detection.limit
}
\description{
Extrapolate vertical profile of a radiometric
quantity (EdZ, EuZ, LuZ) to the air-sea interface using a linear
model applied to log-transformed data in an iterative scheme.
}
\details{
This function is called by
\code{\link{process.EdZ}}
\code{\link{process.EuZ}}
\code{\link{process.LuZ}}
}
\author{
Simon Bélanger
}
