% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop_apstress.R
\name{cop_apstress}
\alias{cop_apstress}
\title{PCOPS version of approximated power stress model.}
\usage{
cop_apstress(
  dis,
  theta = c(1, 1, 1),
  type = "ratio",
  ndim = 2,
  weightmat = 1 - diag(nrow(dis)),
  init = NULL,
  itmaxi = 1000,
  ...,
  stressweight = 1,
  cordweight = 0.5,
  q = 1,
  minpts = ndim + 1,
  epsilon = 10,
  rang = NULL,
  verbose = 0,
  normed = TRUE,
  scale = "sd"
)
}
\arguments{
\item{dis}{numeric matrix or dist object of a matrix of proximities}

\item{theta}{the theta vector of parameters to optimize over. Must be of length three, with the first the kappa argument, the second the lambda argument and the third the nu argument. It can also be a scalar and gets recycled for both ups and tau (so they are equal). One cannot supply upsilon and tau as of yet. Defaults to 1 1 1.}

\item{type}{MDS type.}

\item{ndim}{number of dimensions of the target space}

\item{weightmat}{(optional) a binary matrix of nonnegative weights.}

\item{init}{(optional) initial configuration}

\item{itmaxi}{number of iterations. default is 1000.}

\item{...}{additional arguments to be passed to the fitting procedure}

\item{stressweight}{weight to be used for the fit measure; defaults to 1}

\item{cordweight}{weight to be used for the cordillera; defaults to 0.5}

\item{q}{the norm of the cordillera; defaults to 1}

\item{minpts}{the minimum points to make up a cluster in OPTICS; defaults to ndim+1}

\item{epsilon}{the epsilon parameter of OPTICS, the neighbourhood that is checked; defaults to 10}

\item{rang}{range of the distances (min distance minus max distance). If NULL (default) the cordillera will be normed to each configuration's maximum distance, so an absolute value of goodness-of-clusteredness.}

\item{verbose}{numeric value hat prints information on the fitting process; >2 is extremely verbose}

\item{normed}{should the cordillera be normed; defaults to TRUE}

\item{scale}{should the configuration be scale adjusted}
}
\value{
A list with the components
   \itemize{
        \item{stress:} the stress-1 of the configuration
        \item{stress.m:} default normalized stress (sqrt(stress-1)) 
        \item{copstress:} the weighted loss value
        \item{OC:} the OPTICS cordillera value
        \item{parameters:} the theta parameters used for fitting (kappa, lambda, nu)
        \item{fit:} the returned object of the fitting procedure (typically of class smacofB or smacofP)
        \item{cordillera:} the cordillera object
}
}
\description{
This uses an approximation to power stress that can make use of smacof as workhorse. Free parameters are tau and upsilon.
}
\keyword{multivariate}
