% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cops.R
\name{cop_powerstress}
\alias{cop_powerstress}
\title{COPS version of powerstress}
\usage{
cop_powerstress(dis, theta = c(1, 1, 1), weightmat = 1 - diag(nrow(dis)),
  init = NULL, ndim = 2, ..., stressweight = 1, cordweight = 0.5,
  q = 1, minpts = ndim + 1, epsilon = 10, rang = NULL, verbose = 0,
  scale = 2, normed = TRUE, stresstype = c("default", "stress1",
  "rawstress", "normstress", "enormstress", "enormstress1"))
}
\arguments{
\item{dis}{numeric matrix or dist object of a matrix of proximities}

\item{theta}{the theta vector of powers; the first is kappa (for the fitted distances), the second lambda (for the observed proximities), the third nu (for the weights). If a scalar is given it is recycled.  Defaults to 1 1 1.}

\item{weightmat}{(optional) a matrix of nonnegative weights}

\item{init}{(optional) initial configuration}

\item{ndim}{number of dimensions of the target space}

\item{...}{additional arguments to be passed to the fitting procedure}

\item{stressweight}{weight to be used for the fit measure; defaults to 1}

\item{cordweight}{weight to be used for the cordillera; defaults to 0.5}

\item{q}{the norm of the corrdillera; defaults to 1}

\item{minpts}{the minimum points to make up a cluster in OPTICS; defaults to ndim+1}

\item{epsilon}{the epsilon parameter of OPTICS, the neighbourhood that is checked; defaults to 10}

\item{rang}{range of the distances (min distance minus max distance). If NULL (default) the cordillera will be normed to each configuration's maximum distance, so an absolute value of goodness-of-clusteredness.}

\item{verbose}{numeric value hat prints information on the fitting process; >2 is extremely verbose}

\item{scale}{should the configuration be scaled to mean=0 and sd=1? Defaults to TRUE}

\item{normed}{should the cordillera be normed; defaults to TRUE}

\item{stresstype}{which stress to report? Defaults to explicitly normed stress}
}
\value{
A list with the components
\itemize{
        \item stress: the stress
        \item stress.m: default normalized stress
        \item copstress: the weighted loss value
        \item OC: the Optics cordillera value
        \item parameters: the parameters used for fitting (kappa, lambda)
        \item fit: the returned object of the fitting procedure
        \item cordillera: the cordillera object
}
}
\description{
COPS version of powerstress
}
\keyword{multivariate}
