% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop_rstress.R
\name{cop_rstress}
\alias{cop_rstress}
\title{PCOPS version of rstress}
\usage{
cop_rstress(
  dis,
  theta = 1,
  type = "ratio",
  weightmat = 1 - diag(nrow(dis)),
  init = NULL,
  ndim = 2,
  itmaxi = 10000,
  ...,
  stressweight = 1,
  cordweight = 0.5,
  q = 1,
  minpts = ndim + 1,
  epsilon = 10,
  rang = NULL,
  verbose = 0,
  scale = "sd",
  normed = TRUE
)
}
\arguments{
\item{dis}{numeric matrix or dist object of a matrix of proximities}

\item{theta}{the theta vector of powers; this must be a scalar of the r transformation for the fitted distances proximities. Defaults to 0.5. Note that kappa=2r in powerstress.}

\item{type}{MDS type. Defaults to ratio.}

\item{weightmat}{(optional) a matrix of nonnegative weights}

\item{init}{(optional) initial configuration}

\item{ndim}{number of dimensions of the target space}

\item{itmaxi}{number of iterations. default is 10000.}

\item{...}{additional arguments to be passed to the fitting procedure}

\item{stressweight}{weight to be used for the fit measure; defaults to 1}

\item{cordweight}{weight to be used for the cordillera; defaults to 0.5}

\item{q}{the norm of the cordillera; defaults to 1}

\item{minpts}{the minimum points to make up a cluster in OPTICS; defaults to ndim+1}

\item{epsilon}{the epsilon parameter of OPTICS, the neighbourhood that is checked; defaults to 10}

\item{rang}{range of the distances (min distance minus max distance). If NULL (default) the cordillera will be normed to each configuration's maximum distance, so an absolute value of goodness-of-clusteredness.}

\item{verbose}{numeric value hat prints information on the fitting process; >2 is extremely verbose}

\item{scale}{should the configuration be scale adjusted}

\item{normed}{should the cordillera be normed; defaults to TRUE}
}
\value{
A list with the components
\itemize{
        \item stress: the stress
        \item stress.m: default normalized stress
        \item copstress: the weighted loss value
        \item OC: the Optics cordillera value
        \item parameters: the parameters used for fitting (r)
        \item fit: the returned object of the fitting procedure plus a slot for the original data $deltaorig
        \item cordillera: the cordillera object
}
}
\description{
Free parameter is r for the fitted distances.
}
\keyword{multivariate}
