% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cop_sammon.R
\name{cop_sammon}
\alias{cop_sammon}
\title{PCOPS version of Sammon mapping from MASS}
\usage{
cop_sammon(
  dis,
  theta = 1,
  type = "ratio",
  ndim = 2,
  init = NULL,
  weightmat = NULL,
  itmaxi = 1000,
  ...,
  stressweight = 1,
  cordweight = 0.5,
  q = 1,
  minpts = ndim + 1,
  epsilon = 10,
  rang = NULL,
  verbose = 0,
  scale = "sd",
  normed = TRUE
)
}
\arguments{
\item{dis}{numeric matrix or dist object of a matrix of proximities}

\item{theta}{the theta vector of powers; this must be  a scalar of the lambda transformation for the observed proximities. Defaults to 1.}

\item{type}{MDS type. Ignored here.}

\item{ndim}{number of dimensions of the target space}

\item{init}{(optional) initial configuration}

\item{weightmat}{(optional) a matrix of nonnegative weights.}

\item{itmaxi}{number of iterations. Default is 1000.}

\item{...}{additional arguments to be passed to the fitting procedure}

\item{stressweight}{weight to be used for the fit measure; defaults to 1}

\item{cordweight}{weight to be used for the cordillera; defaults to 0.5}

\item{q}{the norm of the corrdillera; defaults to 1}

\item{minpts}{the minimum points to make up a cluster in OPTICS; defaults to ndim+1}

\item{epsilon}{the epsilon parameter of OPTICS, the neighbourhood that is checked; defaults to 10}

\item{rang}{range of the distances (min distance minus max distance). If NULL (default) the cordillera will be normed to each configuration's maximum distance, so an absolute value of goodness-of-clusteredness.}

\item{verbose}{numeric value hat prints information on the fitting process; >2 is extremely verbose}

\item{scale}{if TRUE the configuration is scale adjusted}

\item{normed}{should the cordillera be normed; defaults to TRUE}
}
\value{
A list with the components
\itemize{
        \item stress: the stress-1
        \item stress.m: default normalized stress (stress-1^2)
        \item copstress: the weighted loss value
        \item OC: the Optics cordillera value
        \item parameters: the parameters used for fitting (lambda)
        \item fit: the returned object of the fitting procedure smacofx::sammon plus a slot for the original data $deltaorig
        \item cordillera: the cordillera object
}
}
\description{
Uses smacofx::sammon wrapper for MASS::sammon. The free parameter is lambda for power transformations of the observed proximities.
}
\keyword{multivariate}
