% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.cops.detection.limit.R
\name{cops.detection.limit}
\alias{cops.detection.limit}
\title{Update the COPS instrument detection limit table}
\usage{
cops.detection.limit(detection.limit)
}
\arguments{
\item{detection.limit}{is a data frame with four column named as
c("waves", "EdZ" ,  "EuZ" ,  "LuZ").}
}
\description{
Update the COPS instrument detection limit table
}
\examples{
 # suppose you want to increase the detection limit of the EdZ 875 channel to 5e-4
df <- detection.limit # get the table from the .GlobalEnv
ix.875 = which(df$waves == 875)
df$EdZ[ix.875] <- 5e-4
cops.detection.limit(df)
# Suppose 565 channel is missing in the
# detection.limit table and you want to add it.
new.channel = c(565, 5e-4, 5e-4, 4e-5)
df<-rbind(df, new.channel) # Add the channel
df<-df[order(df$waves),] # order the data frame
cops.detection.limit(df) # update the Global variable
detection.limit

}
\author{
Simon Belanger
}
