% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_pipeline.R
\name{COPS}
\alias{COPS}
\alias{vertical_pipeline}
\title{Clustering algorithms for Omics based Patient Stratification}
\usage{
COPS(
  dat,
  nfolds = 5,
  nruns = 1,
  association_data = NULL,
  survival_data = NULL,
  module_eigs = NULL,
  verbose = TRUE,
  parallel = 1,
  pathway_enrichment_method = "none",
  multi_omic_methods = NULL,
  vertical_parallelization = FALSE,
  internal_metrics = NULL,
  silhouette_dissimilarity = NULL,
  pre_compute_silhouette_dissimilarity = TRUE,
  ...
)

vertical_pipeline(
  dat_list,
  nfolds = 5,
  nruns = 1,
  survival_data = NULL,
  association_data = NULL,
  multi_omic_methods = NULL,
  parallel = 1,
  data_is_kernels = FALSE,
  silhouette_dissimilarities = NULL,
  by = c("run", "fold", "m", "k", "mkkm_mr_lambda"),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dat}{A single matrix or list of matrices, patients on columns and features on rows.}

\item{nfolds}{Number of cross-validation folds for stability evaluation and metric estimates.}

\item{nruns}{Number of cross-validation replicates for stability evaluation and metric estimates.}

\item{association_data}{Data for association tests, see \code{\link{cluster_associations}} for details.}

\item{survival_data}{Data for survival analysis, see \code{\link{survival_preprocess}} for details.}

\item{module_eigs}{Data for gene module correlation analysis, see \code{\link{gene_module_score}} for details.}

\item{verbose}{Prints progress messages and time taken.}

\item{parallel}{Number of parallel threads for supported operations.}

\item{pathway_enrichment_method}{\code{enrichment_method} for \code{\link{genes_to_pathways}}.}

\item{multi_omic_methods}{Character vector of multi-view clustering method names for \code{\link{multi_omic_clustering}}.}

\item{vertical_parallelization}{(Experimental) if set, all pipeline steps are evaluated in succession within each fold (instead of evaluating each step for all folds before moving on). Always true for multi-view methods.}

\item{internal_metrics}{Internal metric names passed to \code{\link[clusterCrit]{intCriteria}}. This will slow the pipeline down considerably.}

\item{silhouette_dissimilarity}{Dissimilarity matrix to use for computing silhouette indices.}

\item{pre_compute_silhouette_dissimilarity}{If \code{silhouette_dissimilarity=NULL}, it will be calculated automatically by default.}

\item{...}{Extra arguments are passed to pipeline components where appropriate.}

\item{dat_list}{list of data tables}

\item{data_is_kernels}{Whether \code{dat_list} should be treated as kernel matrices}

\item{silhouette_dissimilarities}{list of dissimilarity matrices used for silhouette calculations}

\item{by}{column names used to split threads by}
}
\value{
Returns a \code{list} of pipeline component outputs for each run, fold and 
method given different settings and input data sets. 
\itemize{
  \item{clusters} \code{data.frame} defining clusters
  \item{internal_metrics} \code{data.frame} of internal metrics
  \item{stability} \code{data.frame} of stability scores
  \item{survival} \code{data.frame} of survival analysis results
  \item{modules} \code{data.frame} of gene module association scores
  \item{association} \code{data.frame} of association results to variables of interest
  \item{cluster_sizes} \code{data.frame} giving the sizes of clusters
}

\code{list} of clustering analysis results
}
\description{
Combines \code{\link{subsampling}}, \code{\link{subsample_pathway_enrichment}},
\code{\link{subsample_dimred}}, \code{\link{subsample_clustering_evaluation}}, 
\code{\link{stability_evaluation}}, \code{\link{subsample_survival_evaluation}}, 
\code{\link{subsample_module_evaluation}} and \code{\link{subsample_association_analysis}} 
to conveniently and comprehensively test clustering algorithms on a given set of input data.
}
\details{
If multi_omic_methods is set, then the input matrices are treated as 
different views of the same patients. Available methods are listed in the
documentation for \code{\link{multi_omic_clustering}}.
}
\section{Functions}{
\itemize{
\item \code{vertical_pipeline()}: pipeline vertical parallelization

}}
\examples{
library(COPS)

# Dimensionality reduction and clustering (DR-CL)
res <- COPS(ad_ge_micro_zscore, 
association_data = ad_studies, 
parallel = 1, nruns = 2, nfolds = 5, 
dimred_methods = c("pca", "umap", "tsne"), 
cluster_methods = c("hierarchical", "kmeans"), 
distance_metric = "euclidean", 
n_clusters = 2:4)

# Clustering (CL)
res <- COPS(ad_ge_micro_zscore, 
association_data = ad_studies, 
parallel = 1, nruns = 2, nfolds = 5, 
dimred_methods = c("none"), 
cluster_methods = c("hierarchical"), 
distance_metric = "correlation", 
n_clusters = 2:4)

# Biological knowledge integration and clustering (BK-CL)
res <- COPS(ad_ge_micro_zscore, 
association_data = ad_studies, 
pathway_enrichment_method = "DiffRank", 
gene_key_x = "ENSEMBL", 
gs_subcats = "CP:KEGG_LEGACY", 
parallel = 1, nruns = 2, nfolds = 5, 
dimred_methods = c("none"), 
cluster_methods = c("hierarchical"), 
distance_metric = "correlation", 
n_clusters = 2:4)

}
