% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{cv_association_analysis}
\alias{cv_association_analysis}
\title{Cross-validated cluster association analysis}
\usage{
cv_association_analysis(
  clusters,
  association_data,
  by = c("run", "fold", "datname", "drname", "k", "m"),
  parallel = 1,
  ...
)
}
\arguments{
\item{clusters}{A data.table or data.frame with clustering information.}

\item{association_data}{A data.frame with association variables (categoric or numeric).}

\item{by}{Column names that identify a single clustering result.}

\item{parallel}{Number of parallel threads.}

\item{...}{Extra arguments are ignored.}
}
\value{
\code{data.table} containing association variables
}
\description{
Runs \code{\link{cluster_associations}} in parallel.
}
\details{
Assumes that a parallel backend has been registered for \code{foreach}.
}
