% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{cv_dimred}
\alias{cv_dimred}
\title{Dimensionality reduction on cross-validated data sets}
\usage{
cv_dimred(dat_list, ...)
}
\arguments{
\item{dat_list}{list of data sets}

\item{...}{arguments passed to \code{\link{dim_reduction_suite}}}
}
\value{
list of data sets
}
\description{
Dimensionality reduction on cross-validated data sets
}
