% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{cv_module_evaluation}
\alias{cv_module_evaluation}
\title{Gene module correlation score for multiple clustering results}
\usage{
cv_module_evaluation(
  clusters,
  by = c("run", "fold", "datname", "drname", "k", "m"),
  parallel = 1,
  ...
)
}
\arguments{
\item{clusters}{A data.table or data.frame with clustering information.}

\item{by}{Column names that identify a single clustering result.}

\item{parallel}{Number of parallel threads.}

\item{...}{Extra arguments are passed to \code{\link{gene_module_score}}.}
}
\description{
Runs \code{\link{gene_module_score}} in parallel.
}
\details{
Assumes that a parallel backend has been registered for \code{foreach}.
}
