% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromGenesTo Pathways.R
\name{cv_pathway_enrichment}
\alias{cv_pathway_enrichment}
\title{Pathway enrichment analysis on cross-validated data sets}
\usage{
cv_pathway_enrichment(dat_list, cv_index, gene_id_list, ...)
}
\arguments{
\item{dat_list}{list of data sets}

\item{cv_index}{list of data.frames corresponding to cross-validation fold indicaters as produced by \code{\link[COPS]{cv_fold}}}

\item{gene_id_list}{list of gene name vectors of the corresponding columns in dat_list}

\item{...}{arguments passed on to \code{\link[COPS]{genes_to_pathways}}}
}
\value{

}
\description{
This function wraps the pathway enrichment based data transformations included in this package for use in the pipeline.
}
