% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{data_preprocess}
\alias{data_preprocess}
\title{Data pre-processing for COPS pipelines}
\usage{
data_preprocess(dat, verbose = FALSE)
}
\arguments{
\item{dat}{\code{numeric} \code{matrix} or \code{data.frame} with samples on 
columns, or \code{list} of such datasets}

\item{verbose}{print status messages}
}
\value{
\code{list} containing \code{list} of datasets and \code{list} of 
  feature names (gene ids).
}
\description{
Converts inputs to \code{data.table}s and renames features to \code{"dimX"}.
}
