% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromGenesTo Pathways.R
\name{DiffRank}
\alias{DiffRank}
\title{Expression rank based single sample gene set enrichment analysis}
\usage{
DiffRank(expr, list_db_annots, parallel = 1)
}
\arguments{
\item{expr}{gene expression matrix, samples on columns and genes on rows}

\item{list_db_annots}{list of gene sets with gene names that correspond to rows in \strong{expr}}

\item{parallel}{a numeric value indicating the number of processors to use when doing the calculations in parallel.}
}
\value{

}
\description{
DiffRank by Wang et al. BMC Medical Genomics 2019
}
