% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_cls_pipeline.R
\name{dimred_cluster}
\alias{dimred_cluster}
\title{Reduce dimensionality and cluster input data}
\usage{
dimred_cluster(
  dat_list,
  best = NULL,
  data_id = 1,
  dimred_method = "pca",
  dimred_dim = 2,
  dimred_perp = 30,
  cluster_method = "kmeans",
  k = 2,
  metric = "euclidean"
)
}
\arguments{
\item{dat_list}{\code{list} of pre-processed data sets, original input to pipeline or
containing element \code{data_id} corresponding to the preferred input}

\item{best}{a single row \code{data.frame} which specifies the preferred methods}

\item{data_id}{name or index of preferred element of \code{dat_list}}

\item{dimred_method}{name of dimensionality reduction method for
\code{\link{dim_reduction_suite}}}

\item{dimred_dim}{number of dimensions parameter for \code{\link{dim_reduction_suite}}}

\item{dimred_perp}{perplexity parameter for \code{\link{dim_reduction_suite}}}

\item{cluster_method}{clustering method name}

\item{k}{number of clusters parameter}

\item{metric}{distance metric parameter for clustering}
}
\value{
Returns a \code{list} with \code{$embedding} corresponding to the embedding
        generated using \code{\link{dim_reduction_suite}} and \code{$clustering}
        corresponding to clustering on \code{$embedding} using
        \code{\link[clValid]{clValid}}.
}
\description{
Convenience function that runs a specific setting of the dimensionality reduction
and clustering analysis pipeline. Accepts \code{\link{clusteval_scoring}} output
\code{$best} as input.
}
