% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_cls_pipeline.R
\name{dimred_clusteval_pipeline}
\alias{dimred_clusteval_pipeline}
\title{Pipeline combining dimensionality reduction, clustering, cross-validation and evaluation}
\usage{
dimred_clusteval_pipeline(
  dat_list,
  batch_label = NULL,
  verbose = TRUE,
  parallel = 1,
  dim_reduction = TRUE,
  pre_clust_cv = FALSE,
  ...
)
}
\arguments{
\item{dat_list}{list of pre-processed data sets to run pipeline for}

\item{batch_label}{vector or matrix with categorical variables on columns}

\item{verbose}{if \code{TRUE}, prints progress notifications}

\item{parallel}{sets up and registers \code{parallel} number of threads for supported operations}

\item{dim_reduction}{set to \code{FALSE} to skip the dimensionality reduction}

\item{pre_clust_cv}{set to \code{TRUE} if data is already organized into folds from previous step}

\item{...}{extra arguments are passed to pipeline components where appropriate}
}
\value{
Returns a \code{list} of pipeline component outputs for given settings and input data sets
}
\description{
Combines \code{\link{dim_reduction_suite}}, \code{\link{clusteval_cv}} and
\code{\link{stability_eval}}.
}
