% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{DSC}
\alias{DSC}
\title{Dispersion Separability Criterion}
\usage{
DSC(data_matrix, batch_label)
}
\arguments{
\item{data_matrix}{numeric matrix, samples on columns}

\item{batch_label}{categorical variable, must be vector}
}
\value{
Returns the DSC of \code{data_matrix} with respect to \code{batch_label} as a scalar value
}
\description{
Used by TCGA Batch Effects Viewer \url{https://bioinformatics.mdanderson.org/public-software/tcga-batch-effects/}.
Based on \url{http://citeseerx.ist.psu.edu/viewdoc/summary?doi=10.1.1.95.1128}.
Can be used to measure batch effect. TODO: fix implementation
}
