% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromGenesTo Pathways.R
\name{ecdf_transform}
\alias{ecdf_transform}
\title{Empirical cumulative density function transformation}
\usage{
ecdf_transform(x, parallel = 1)
}
\arguments{
\item{x}{numerical matrix}

\item{parallel}{number of threads}
}
\value{
matrix of row-wise ecdf values with dimensions matching input
}
\description{
Estimate each feature (row) distribution using Gaussian kernels with sigma corresponding to sd.
}
