% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_kernel.R
\name{ECMC}
\alias{ECMC}
\title{Enhanced consensus multi-view clustering}
\usage{
ECMC(x, a, b, eps = 1e-06, maxiter = 10, parallel = 0)
}
\arguments{
\item{x}{list of kernel matrices}

\item{a}{consensus reward}

\item{b}{disagreement penalty}

\item{eps}{stopping threshold for mean difference in solution}

\item{maxiter}{maximum number of iterations before terminating}

\item{parallel}{number of threads for optimization}
}
\value{
list of kernel matrices, disagreement matrices and solution diagnostics
}
\description{
Consensus kernel approach described by Cai & Li (2017).
}
