% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{embedding_associations}
\alias{embedding_associations}
\title{Continuous variable to categorical variable association analysis}
\usage{
embedding_associations(dat, class, n_pc_max = 10, ...)
}
\arguments{
\item{dat}{data matrix, samples on columns}

\item{class}{vector or matrix where columns are categorical variables}

\item{n_pc_max}{maximum number of principal components to analyze}

\item{...}{extra arguments are ignored}
}
\value{
Returns a \code{list} containing \code{\link[kBET]{batch_sil}},
        \code{\link[kBET]{pcRegression}} and \code{\link{DSC}}
        computed for each column of \code{class}
}
\description{
Three non-clustering batch effect estimators put together. \cr
Can be used for other purposes as well.
}
