% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etirr.R
\name{etirr}
\alias{etirr}
\title{Extraterrestrial Solar Irradiance (F0) from Thuillier}
\usage{
etirr(waves)
}
\arguments{
\item{waves}{is a vector for wavelengths}
}
\value{
Returns F0 values in W/m^2/nm
}
\description{
Extraterrestrial Solar Irradiance (F0) from Thuillier
}
\author{
Bernard Gentilly
}
