% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{expressionToRWFeatures}
\alias{expressionToRWFeatures}
\title{Gene expression to gene seeds to RWR on PPI into FGSEA scores}
\usage{
expressionToRWFeatures(
  dat,
  disease_id,
  otp_score = "association_score.datatypes.rna_expression",
  otp_cutoff = 0,
  ppi_cutoff = 700,
  pw_min.size = 5,
  pw_max.size = 200,
  dat_gene_key = "SYMBOL",
  gs_subcats = c("BP", "MF", "CP:KEGG_LEGACY", "CP:REACTOME"),
  directed_ppi = FALSE,
  ...
)
}
\arguments{
\item{dat}{A gene expression matrix with samples on columns.}

\item{disease_id}{Integer ID in Open Targets platform.}

\item{otp_score}{Name of association score column in Open Targets.}

\item{otp_cutoff}{Numeric association score cutoff for Open Targets platform genes.}

\item{ppi_cutoff}{Numeric PPI link score cutoff.}

\item{pw_min.size}{Minimum gene set size to use.}

\item{pw_max.size}{Maximum gene set size to use.}

\item{dat_gene_key}{Data gene annotation type.}

\item{gs_subcats}{Gene set subcategories in \code{\link[msigdbr]{msigdbr}} to retrieve for enrichment analysis.}

\item{directed_ppi}{Whether to generate a directed network.}

\item{...}{extra arguments are passed on to \code{\link{RWRFGSEA}}}
}
\value{
list of enrichment scores
}
\description{
Conveniently wraps \code{\link{RWRFGSEA}} using 
\code{\link{retrieveDiseaseGenesOT}} for disease associated genes, 
\code{STRINGdb} human PPI as a network, \code{msigdbr} for gene set annotations and 
\code{biomaRt} for matching gene IDs.
}
