% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{feature_associations}
\alias{feature_associations}
\title{Feature batch-effect analysis}
\usage{
feature_associations(x, class, n_pc_max = 10, ...)
}
\arguments{
\item{x}{data matrix, samples on columns}

\item{class}{vector, data.frame or list where columns are categorical variables}

\item{n_pc_max}{maximum number of principal components to analyze}

\item{...}{extra arguments are ignored}
}
\value{
Returns a \code{list} containing the following objects: 
\itemize{
  \item{associations}{\code{data.frame} summary of \code{\link[kBET]{batch_sil}},
                      \code{\link[kBET]{pcRegression}} (\code{$maxR2}) and 
                      \code{\link{DSC}} computed for each column of \code{class}}
  \item{eigenvalues}{PC eigenvalues and percentage of explained variance from 
                     \code{\link[FactoMineR]{PCA}}}
}
}
\description{
Three batch effect estimators for numeric features put together. \cr
Can be used for other purposes, e.g., as a multivariate association index.
}
