% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromGenesTo Pathways.R
\name{fgsea_wrapper}
\alias{fgsea_wrapper}
\title{FGSEA wrapper}
\usage{
fgsea_wrapper(data_matrix, list_db_annots, value_cutoff = 0, parallel = 1, ...)
}
\arguments{
\item{data_matrix}{numeric values used for ranking, assumes samples on columns}

\item{list_db_annots}{list of gene sets or pathway annotations}

\item{value_cutoff}{cutoff for inclusion in analysis}

\item{parallel}{}

\item{...}{}
}
\value{

}
\description{
Wrapper for fast pre-ranked gene set enrichment analysis (\code{\link[fgsea]{fgseaSimple}})
}
