% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{format_scores}
\alias{format_scores}
\title{Renames internal method and score names to something more understandable.}
\usage{
format_scores(x, multi_omic = FALSE)
}
\arguments{
\item{x}{scores from COPS pipeline}
}
\description{
Renames internal method and score names to something more understandable.
}
