% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromGenesTo Pathways.R
\name{fromGeneToNetworksToPathwayFeatures}
\alias{fromGeneToNetworksToPathwayFeatures}
\title{Random walk with restart and FGSEA worker}
\usage{
fromGeneToNetworksToPathwayFeatures(
  dat,
  disease.genes,
  gene.network,
  list_db_annots,
  top.ranked.genes = 100,
  parallel = 1,
  verbose = FALSE,
  rwr_restart = 0.33,
  rwr_norm = "none",
  rwr_affinity_norm = "quantile",
  rwr_cutoff = 0,
  kernelized_sorting = FALSE,
  reg_up_down = FALSE,
  ...
)
}
\arguments{
\item{dat}{a gene expression matrix with samples on columns}

\item{disease.genes}{a character vector containing Entrez IDs of genes associated with the target disease}

\item{gene.network}{an \code{igraph.object} with nodes matching to \code{dat} rows}

\item{list_db_annots}{list of character vectors containing gene sets for FGSEA}

\item{top.ranked.genes}{integer, controls the number of gene seed candidates to intersect with the disease genes}

\item{parallel}{integer, number of threads}

\item{verbose}{boolean, verbosity of \code{dnet::dRWR}}

\item{rwr_restart}{the restart probability used for RWR. See \code{dnet::dRWR} for more details.}

\item{rwr_norm}{the way to normalise the adjacency matrix of the input graph. See \code{dnet::dRWR} for more details.}

\item{rwr_cutoff}{the cuoff value to select the most visited genes.}

\item{...}{extra arguments are ignored}

\item{min.size}{integer, minimum size of gene sets}

\item{max.size}{integer, maximum size of gene sets}
}
\value{
list of enrichment scores
}
\description{
Runs random walk in a given network starting from seed genes selected from most over/under expressed in the data intersected with 
known disease genes. The gene affinities are then processed with FGSEA to yield pathway features.
}
