% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromGenesTo Pathways.R
\name{fromGeneToPathwayFeatures}
\alias{fromGeneToPathwayFeatures}
\title{Transform a gene-level data matrix into path-level information}
\usage{
fromGeneToPathwayFeatures(
  dat,
  study_batch = NULL,
  min.size = 5,
  max.size = 200,
  parallel = 4,
  verbose = FALSE,
  key_name = "ENSEMBL",
  kcdf = "Gaussian",
  rnaseq = FALSE
)
}
\arguments{
\item{dat}{a numeric matrix representing gene expression profiles. Genes on the rows and samples on the columns.}

\item{study_batch}{factor indicating the individual studies. 
A NULL value implicates that the input data matrix represents the original dataset (without removing the batches).}

\item{min.size}{a numeric value indicating the minimum size of the resulting gene sets.}

\item{max.size}{a numeric value indicating the maximum size of the resulting gene sets.}

\item{parallel}{a numeric value indicating the umber of processors to use when doing the calculations in parallel.}

\item{verbose}{controls verbosity}

\item{kcdf}{distribution name for \code{\link[GSVA]{gsva}} empirical distribution kernel}
}
\value{
a list of three data frames:\cr
       - \strong{KEGG_PW}: a character variable containing entrez gene ids;\cr
       - \strong{GO_PW}: a character variable containing gene symbols;\cr
       - \strong{REACTOME_PW}: a numeric variable containing gene-disease scores.
}
\description{
Utility function to extract pathway-based features from gene expression data.
}
