% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{gene_module_eigengenes}
\alias{gene_module_eigengenes}
\title{Gene module eigen genes}
\usage{
gene_module_eigengenes(x, modules, npc = 1)
}
\arguments{
\item{x}{Gene expression data with samples on columns.}

\item{modules}{Gene clustering result.}

\item{npc}{Number of PCs to extract per module.}
}
\description{
Computes module eigen genes by applying PCA to gene modules
}
