% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.cops.DB.R
\name{generate.cops.DB}
\alias{generate.cops.DB}
\title{Generate an AOPs data base derived from COPS light profiles for a list of directories}
\usage{
generate.cops.DB(
  path = "./",
  waves.DB = c(320, 330, 340, 380, 412, 443, 465, 490, 510, 532, 555, 589, 625, 665,
    683, 694, 710, 780, 875),
  mission = "XXX"
)
}
\arguments{
\item{path}{is the path where the file directories.for.cops.dat containing the data folders
to merge in the data base.}

\item{mission}{is a string for the name of the mission. It will be used for the file names of the output.}

\item{wave.DB}{is a vector of wavelengths to include un the data base. Default is
waves.DB=c(305, 320, 330, 340,380, 412, 443, 465,490, 510, 532, 555,589, 625, 665, 683, 694, 710, 780)}
}
\value{
It returns a list object named COPS.DB containing matrices of
mean and standard deviation
of Kd1p, Kd10p, Rrs, Ed0.0p,  Ed0.f and vectors for
date, lat, lon sunzen and waves

The object COPS.DB is saved in RData format. The data are also saved in ASCII (.dat with comma separator)
and a figure showing the measured rho_w spectra of the data base is produced.
}
\description{
The AOPs (Kd_s, Rrs) derived from valid profiles are averaged.
IMPORTANT: the only the cast selected in select.cops.dat will be average.
In addition, the extrapolation method for Rrs.0p must be specified in the
third column of the select.cops.dat.
}
