% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment.R
\name{genes_to_pathways}
\alias{genes_to_pathways}
\alias{DiffRank}
\title{Transform a gene-level data matrix into path-level information}
\usage{
genes_to_pathways(
  x,
  enrichment_method = "GSVA",
  gene_set_list = NULL,
  batch_label_pw = NULL,
  min_size = 5,
  max_size = 200,
  parallel = 1,
  verbose = FALSE,
  rwrfgsea_verbose = FALSE,
  gene_key_x = "SYMBOL",
  gs_subcats = c("GO:BP", "GO:MF", "CP:KEGG", "CP:REACTOME"),
  gsva_kcdf = "Gaussian",
  ...
)

DiffRank(x, gene_set_list, parallel = 1)
}
\arguments{
\item{x}{gene feature matrix, samples on columns and genes on rows.}

\item{enrichment_method}{options: "DiffRank", "GSVA", "RWRFGSEA"}

\item{gene_set_list}{list of gene sets with gene names that correspond to rows in \strong{x}}

\item{batch_label_pw}{batch labels for batch-wise enrichment, normal enrichment performed if NULL}

\item{min_size}{a numeric value indicating the minimum size of gene sets included}

\item{max_size}{a numeric value indicating the maximum size of gene sets included}

\item{parallel}{a numeric value indicating the number of processors to use when doing the calculations in parallel.}

\item{verbose}{controls verbosity}

\item{rwrfgsea_verbose}{controls \code{\link[COPS]{RWRFGSEA}} verbosity which can be a lot when running single-threaded}

\item{gene_key_x}{if \code{is.null(gene_set_list)} and \code{x} rownames are not gene symbols, this specifies the column name in 
\code{\link[org.Hs.eg.db]{org.Hs.eg.db}} to translate pathway gene symbols to. The default value results in gene symbol based gene 
sets when \code{is.null(gene_set_list)}.}

\item{gs_subcats}{if \code{is.null(gene_set_list)}, this character vector indicates \code{\link[msigdbr]{msigdbr}}) gene set subcollection 
names that are included in the analysis.}

\item{gsva_kcdf}{distribution name for \code{\link[GSVA]{gsva}} empirical probability distribution kernel.}

\item{...}{extra arguments are passed to \code{\link[COPS]{RWRFGSEA}}}
}
\value{
a list of \code{data.frame}s corresponding to the transformed features 
  based on the selected gene sets (only supports GO, KEGG and REACTOME at the moment)

\code{data.frame} of pathway activities for each sample
}
\description{
This is a utility function which wraps several methods to extract pathway-based features from gene feature data
}
\section{Functions}{
\itemize{
\item \code{DiffRank()}: DiffRank by Wang et al. BMC Medical Genomics 2019

}}
\examples{
library(COPS)

## GSVA example
ad_gsva <- genes_to_pathways(
    ad_ge_micro_zscore, 
    "GSVA", 
    parallel = 1, 
    gene_key_x = "ENSEMBL", 
    gs_subcats = "CP:KEGG_LEGACY")
# batch-wise
ad_gsva <- genes_to_pathways(
    ad_ge_micro_zscore, 
    "GSVA", 
    batch_label_pw = ad_studies$GSE, 
    parallel = 1, 
    gene_key_x = "ENSEMBL", 
    gs_subcats = "CP:KEGG_LEGACY")

## DiffRank example
ad_diffrank <- genes_to_pathways(
    ad_ge_micro_zscore, 
    "DiffRank", 
    parallel = 1, 
    gene_key_x = "ENSEMBL", 
    gs_subcats = "CP:KEGG_LEGACY")
# batch-wise
ad_diffrank <- genes_to_pathways(
    ad_ge_micro_zscore, 
    "DiffRank", 
    batch_label_pw = ad_studies$GSE, 
    parallel = 1, 
    gene_key_x = "ENSEMBL", 
    gs_subcats = "CP:KEGG_LEGACY")

}
