% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_kernel_weights}
\alias{get_kernel_weights}
\title{Extract kernel weights from MKKM-MR clustering results}
\usage{
get_kernel_weights(clusters)
}
\arguments{
\item{clusters}{output from \code{\link{multi_omic_clustering}} run with mkkm_mr}
}
\value{
\code{data.frame} of kernel weights for each MKKM-MR result
}
\description{
Extract kernel weights from MKKM-MR clustering results
}
