% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getHumanPPIfromSTRINGdb}
\alias{getHumanPPIfromSTRINGdb}
\title{Retrieve human protein-protein interaction network from STRINGdb.}
\usage{
getHumanPPIfromSTRINGdb(
  gene_ids = NULL,
  cutoff = 700,
  directed = FALSE,
  version = "11",
  gene_id_mart_column = "hgnc_symbol"
)
}
\arguments{
\item{gene_ids}{a character vector indicating target genes to include.}

\item{cutoff}{a numeric value indicating the cutoff for the edge scores.}

\item{directed}{a boolean value indicating the type of grpah to be generated.}

\item{version}{a character value specifying STRINGdb version to query from}
}
\value{
an igraph object.
}
\description{
Utility function to extract a gene subnetwork from STRINGdb including only the seed genes and their interactions.
}
