% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromGenesTo Pathways.R
\name{GSVA}
\alias{GSVA}
\title{Transform a gene-level data matrix into path-level information}
\usage{
GSVA(dat, list_db_annots, kcdf = "Gaussian", parallel = 1, ...)
}
\arguments{
\item{dat}{a numeric matrix representing gene expression profiles. Genes on the rows and samples on the columns.}

\item{kcdf}{distribution name for \code{\link[GSVA]{gsva}} empirical distribution kernel}

\item{parallel}{a numeric value indicating the number of processors to use when doing the calculations in parallel.}

\item{study_batch}{factor indicating the individual studies. 
A NULL value implicates that the input data matrix represents the original dataset (without removing the batches).}

\item{min.size}{a numeric value indicating the minimum size of gene sets included}

\item{max.size}{a numeric value indicating the maximum size of gene sets included}

\item{verbose}{controls verbosity}

\item{gs_subcats}{a character vector indicating msigdbr gene set subcategory names to include in the analysis}
}
\value{
a list of three data frames:\cr
       - \strong{KEGG_PW}: a character variable containing entrez gene ids;\cr
       - \strong{GO_PW}: a character variable containing gene symbols;\cr
       - \strong{REACTOME_PW}: a numeric variable containing gene-disease scores.
}
\description{
Utility function to extract pathway-based features from gene expression data using GSVA.
}
