% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterization.R
\name{heatmap_annotations}
\alias{heatmap_annotations}
\title{Heatmap annotation automation}
\usage{
heatmap_annotations(
  annotations,
  factor_color_sets = list(RColorBrewer::brewer.pal(8, "Dark2"),
    RColorBrewer::brewer.pal(9, "Set1"), RColorBrewer::brewer.pal(9, "Pastel1"),
    pals::watlington(n = 16), pals::kelly(n = 22)),
  ...
)
}
\arguments{
\item{annotations}{list of variables}

\item{factor_color_sets}{list of color sets (not in order)}

\item{...}{passed onto \code{\link[ComplexHeatmap]{HeatmapAnnotation}}}
}
\value{
\code{\link[ComplexHeatmap]{`HeatmapList-class`}}
}
\description{
Generates stacked \code{\link[ComplexHeatmap]{HeatmapAnnotation}}s from a 
list of annotations.
}
\details{
Characters and factors are assigned colors from categorical palettes 
according to the number of categories. NOTE: if a palette does not have 
enough colors, the colors are assigned from the next palette in the list.
}
