% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{jaccard_matrix}
\alias{jaccard_matrix}
\title{Jaccard index between indicator matrix columns}
\usage{
jaccard_matrix(x)
}
\arguments{
\item{x}{indicator or binary feature matrix}
}
\value{
\code{matrix}
}
\description{
Jaccard index between indicator matrix columns
}
