% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{JaccardSimCoef}
\alias{JaccardSimCoef}
\title{Jaccard similarity coefficient between two categorical vectors}
\usage{
JaccardSimCoef(a, b)
}
\arguments{
\item{a}{categorical vector}

\item{b}{categorical vector}
}
\value{
Scalar, Jaccard coefficient
}
\description{
R implementation of clusteval::cluster_similarity to avoid crashing.
}
