% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{jdist}
\alias{jdist}
\title{Average Jaccard dissimilarity coefficient between multiple clustering results}
\usage{
jdist(clustering_list)
}
\arguments{
\item{clustering_list}{\code{list} of clustering vectors}
}
\value{
Returns a scalar value giving the mean jaccard distance between clustering vectors
}
\description{
Measures the stability of a given clustering method when combined with
cross-validation or bootstrap by estimating the repeatability of clustering results.
}
