% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{jdist_ref}
\alias{jdist_ref}
\title{Average Jaccard dissimilarity coefficient between clusterings and references}
\usage{
jdist_ref(clustering_list, clustering_reference_list)
}
\arguments{
\item{clustering_list}{\code{list} of clustering vectors}

\item{clustering_reference_list}{\code{list} of reference clustering vectors}
}
\value{
Returns a scalar value giving the mean jaccard distance between clustering vectors
}
\description{
Measures the stability of a given clustering method when combined with
cross-validation or bootstrap by estimating the repeatability of clustering results.
}
