% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kernelized_kmeans}
\alias{kernelized_kmeans}
\title{Kernelized k-means}
\usage{
kernelized_kmeans(K, n_k, num_init = 100, max_iter = 100, parallel = 1, ...)
}
\arguments{
\item{K}{kernel \code{matrix}}

\item{n_k}{number of clusters}

\item{num_init}{number of initializations}

\item{max_iter}{maximum number of k-means iterations}

\item{parallel}{number of parallel threads for running different initializations}

\item{...}{ignored}
}
\value{
\code{list} of cluster assignments and k-means objective
}
\description{
K-means++ without spectral relaxation.
}
