% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_clustering.R
\name{knn_communities}
\alias{knn_communities}
\title{Title}
\usage{
knn_communities(dat, k = 30, jaccard_kernel = TRUE)
}
\arguments{
\item{dat}{}

\item{k}{}

\item{jaccard_kernel}{}
}
\value{

}
\description{
Title
}
