% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchphi-data.R
\name{matchphi}
\alias{matchphi}
\title{Distances of MATCH-ADTC modules}
\format{
A 32 x 32 distance matrix.
}
\description{
A symmetric distance matrix of 32 MATCH-ADT modules based on their usage by clinicians. The raw data were counts of how often each module has been used with each of 449 youths, resulting in a count profile of each module. Based on that the phi-distance between the modules has been calculated.
}
\details{
This is an object that inherits from class distance (see package analogue) and matrix.
}
