% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerstress.R
\name{mkPowerALTERN}
\alias{mkPowerALTERN}
\title{Take matrix to a power}
\usage{
mkPowerALTERN(x, r)
}
\arguments{
\item{x}{matrix}

\item{r}{numeric (power)}
}
\value{
a matrix
}
\description{
Take matrix to a power
}
