% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{module_evaluation}
\alias{module_evaluation}
\title{Gene module correlation score for multiple clustering results}
\usage{
module_evaluation(
  clusters,
  module_eigs,
  threshold = 0.3,
  nan.substitute = 0,
  by = c("run", "fold", "datname", "drname", "k", "m"),
  ...
)
}
\arguments{
\item{clusters}{A data.table or data.frame with clustering information.}

\item{module_eigs}{See \code{\link{gene_module_score}}.}

\item{threshold}{See \code{\link{gene_module_score}}.}

\item{nan.substitute}{See \code{\link{gene_module_score}}.}

\item{by}{Column names that identify a single clustering result.}

\item{...}{Extra arguments are ignored.}
}
\value{

}
\description{
Runs \code{\link{gene_module_score}} in parallel.
}
\details{
Assumes that a parallel backend has been registered for \code{foreach}.
}
