% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.plot.Rrs.Kd.for.station.R
\name{new.plot.Rrs.Kd.for.station}
\alias{new.plot.Rrs.Kd.for.station}
\title{This function plot the replicates of Rrs and
the integrated Kd (K0) from surface to a given depth.}
\usage{
new.plot.Rrs.Kd.for.station(path = "./", depthEdZ = NA)
}
\arguments{
\item{path}{is the path where raw COPS data are located.}

\item{depthEdZ}{is the depth of the ingetration of Kd.
Default is NA and the code will take the same depth used for
the linear interpolation.}
}
\value{
It produces 2 plots in PNG format in the working directory.
}
\description{
This function plot the replicates of Rrs and
the integrated Kd (K0) from surface to a given depth.
}
\author{
Simon Belanger, UQAR

Charles-André Roux
}
