% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{node_betweenness_parallel}
\alias{node_betweenness_parallel}
\title{Parallel computation of node betweenness}
\usage{
node_betweenness_parallel(
  networks,
  parallel = 1,
  pathway_node_betweenness_endpoints = TRUE,
  pathway_first_shortest_path = FALSE
)
}
\arguments{
\item{networks}{\code{list} of \code{igraph} objects}

\item{parallel}{number of threads}
}
\value{
\code{list} of betweenness centrality vectors
}
\description{
Parallel computation of node betweenness
}
