% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{normalize_kernel}
\alias{normalize_kernel}
\title{Kernel normalization to unit norm}
\usage{
normalize_kernel(x)
}
\arguments{
\item{x}{kernel matrix}
}
\value{
normalized kernel matrix
}
\description{
Kernel normalization to unit norm
}
