% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orbex.R
\name{orbex}
\alias{orbex}
\title{Correction factor for Sun-Earth distance variation thoughout the year}
\usage{
orbex(month, day)
}
\arguments{
\item{month}{is the month (1 to 12)}

\item{day}{is the day of the month (1-31)}
}
\value{
Returns a factor that multiply F0 to account for Sun-Earth distance variation at any given date.
This factor can change the irradiance up to 3.3\\%.
}
\description{
Correction factor for Sun-Earth distance variation thoughout the year
}
\author{
Bernard Gentili
}
