% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{pamogk}
\alias{pamogk}
\alias{node_betweenness_parallel}
\alias{weighted_linear_kernel}
\title{PAMOGK (Tepeli et al. 2021)}
\usage{
pamogk(
  x,
  networks,
  shortest_paths,
  z_up = qnorm(0.975),
  z_down = qnorm(0.025),
  gene_key = "SYMBOL",
  parallel = 1
)

node_betweenness_parallel(networks, parallel = 1)

weighted_linear_kernel(x, weights)
}
\arguments{
\item{x}{feature matrix}

\item{networks}{list of igraph objects}

\item{shortest_paths}{}

\item{z_up}{}

\item{z_down}{}

\item{gene_key}{}

\item{parallel}{number of threads}

\item{weights}{feature weights}
}
\value{




kernel matrix
}
\description{
PAMOGK (Tepeli et al. 2021)
}
\section{Functions}{
\itemize{
\item \code{node_betweenness_parallel}: Parallel computation of node betweenness

\item \code{weighted_linear_kernel}: Weighted linear kernel
}}

