% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pareto_plot}
\alias{pareto_plot}
\title{Pairwise plots for visual Pareto based multi-objective optimization}
\usage{
pareto_plot(
  scores,
  plot_palette = pals::watlington(16),
  metrics = c("TrainStabilityJaccard", "Silhouette", "Smallest_cluster_size"),
  metrics_scale = rep("identity", length(metrics)),
  color_var = "Transform",
  shape_var = "Clustering",
  size_var = "k",
  size_range = c(2, 6),
  color_scale = ggplot2::scale_color_manual(values = plot_palette)
)
}
\arguments{
\item{scores}{\code{data.frame} of scores}

\item{plot_palette}{color codes used for coloring based on \code{color_var}}

\item{metrics}{column names in \code{scores} to plot}

\item{color_var}{column name in \code{scores} for color}

\item{shape_var}{column name in \code{scores} for shape}

\item{size_var}{column name in \code{scores} for size}
}
\value{
\code{\link[gridExtra]{grid.arrange}} result
}
\description{
Pairwise plots for visual Pareto based multi-objective optimization
}
