% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{pathway_gene_subnetworks}
\alias{pathway_gene_subnetworks}
\title{Extract subnetworks based on pathway genesets}
\usage{
pathway_gene_subnetworks(gene_network, gene_sets)
}
\arguments{
\item{gene_network}{\code{igraph} object}

\item{gene_sets}{\code{list} of gene sets}
}
\value{
list of kernel matrices
}
\description{
Extract subnetworks based on pathway genesets
}
\examples{
\dontrun{
pw_db <- msigdbr::msigdbr(
    species = "Homo sapiens", 
    category = "C2", 
    subcollection = "CP:KEGG_LEGACY"
)
pw_list <- lapply(split(pw_db, pw_db$gs_name), function(x) x$ensembl_gene)
pw_list <- pw_list[which(sapply(pw_list, length) <= 200 & sapply(pw_list, length) >= 5)]
ppi_net <- COPS::getHumanPPIfromSTRINGdb(gene_id_mart_column = "ensembl_gene_id")
ppi_pws <- COPS::pathway_gene_subnetworks(ppi_net, pw_list)
# Check the connected components to see if subnets are properly connected:
lapply(ppi_pws, function(x) igraph::components(x)$csize) # Not all are
}

}
