% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brca-data.R
\docType{data}
\name{pathway_networks}
\alias{pathway_networks}
\title{Pathway networks associated with cancer hallmarks}
\format{
An object of class \code{list} containing \code{igraph} objects.
}
\usage{
data(pathway_networks)
}
\description{
Associations from Zhang et al. 2020 (https://doi.org/10.3389/fgene.2020.00029). 
Networks from the KEGG pathway database (https://doi.org/10.1093/nar/gkac963).
}
\keyword{datasets}
