% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pca_viz}
\alias{pca_viz}
\title{Data visualization using PCA}
\usage{
pca_viz(data, category, category_label)
}
\arguments{
\item{data}{}

\item{category}{}

\item{category_label}{}
}
\value{

}
\description{
Data visualization using PCA
}
