% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerstress.R
\name{pdist}
\alias{pdist}
\title{Squared p-distances}
\usage{
pdist(x, p)
}
\arguments{
\item{x}{numeric matrix}

\item{p}{p>0 the Minkoswki distance}
}
\value{
squared Minkowski distance matrix
}
\description{
Squared p-distances
}
