% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phidist.R
\name{phidistance}
\alias{phidistance}
\title{Calculating the pairwise phi distance matrix between n vectors}
\usage{
phidistance(X)
}
\arguments{
\item{X}{an n times p numeric matrix or data frame}
}
\value{
a symmetric n times n matrix of pairwise phi distance (between rows of X) with 0 in the main diagonal. Is an object of class distance and matrix.
}
\description{
The pairwise phi-distance of two vectors x and y is sqrt(sum(((x[i]-y[i])^2)/((x[i]+y[i])*(sum(x)+sum(y))))). The function calculates this for all pairs of rows of a matrix or data frame X.
}
