% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{PIK}
\alias{PIK}
\alias{PIK_KEGG}
\alias{KEGG_networks}
\alias{PIK_from_networks}
\title{Pathway induced kernel (Manica et al. 2019)}
\usage{
PIK(x, L, rwr_smoothing = FALSE, ...)

PIK_KEGG(x, gene_key = "SYMBOL", ...)

KEGG_networks()

PIK_from_networks(x, networks, normalized_laplacian = TRUE, parallel = 1, ...)
}
\arguments{
\item{x}{gene feature matrix}

\item{L}{pathway graph Laplacian matrix}

\item{rwr_smoothing}{apply feature smoothing}

\item{...}{}

\item{gene_key}{}

\item{networks}{list of igraph objects corresponding to pathway graphs}

\item{normalized_laplacian}{}

\item{parallel}{}
}
\value{
kernel matrix

list of undirected igraph objects

list of kernel matrices
}
\description{
Pathway induced kernel (Manica et al. 2019)
}
\section{Functions}{
\itemize{
\item \code{PIK_KEGG()}: Use KEGG pathways to form pathway induced kernels

\item \code{KEGG_networks()}: Download KEGG pathway graphs

\item \code{PIK_from_networks()}: Pathway induced kernel from networks

}}
