% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{PIK_KEGG}
\alias{PIK_KEGG}
\title{KEGG pathway induced kernels}
\usage{
PIK_KEGG(x, normalized_laplacian = TRUE, gene_key = "SYMBOL")
}
\arguments{
\item{x}{gene expression matrix}

\item{normalized_laplacian}{}

\item{gene_key}{}
}
\value{

}
\description{
KEGG pathway induced kernels
}
