% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.all.bottom.spectra.R
\name{plot.all.bottom.spectra}
\alias{plot.all.bottom.spectra}
\title{This function finds all cast .RData files in the station
subdirectories, and calls a plotting function on each of them.}
\usage{
\method{plot}{all.bottom.spectra}(functionName, save = F)
}
\arguments{
\item{functionName}{is an optional argument specifying the
name of the plotting function to be used. plot.bottom.spectra() is called
by default}

\item{save}{is an optional argument specifying whether the
generated plots should be saved in a folder called
"bottom_spectra" in the working directoy}
}
\description{
This function must be launched from the L2 directory, which
should contain all processed data folders containing "Station"
in their name.
}
\details{
The plotting function must have a parameter SAVE (boolean)
}
\author{
Charles-André Roux
}
