% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.all.Kd.Rrs.R
\name{plot.all.Kd.Rrs}
\alias{plot.all.Kd.Rrs}
\title{This function finds and calls a specified plotting function in all valid directories}
\usage{
\method{plot}{all.Kd.Rrs}(functionName = "plot.Rrs.Kd.for.station")
}
\arguments{
\item{functionName}{is an optional argument specifying the
name of the plotting function to be used. new.plot.Rrs.Kd.for.station is called
by default}
}
\description{
This function must be launched from the L2 directory, which
should contain all processed data folders containing "Station"
in their name.
}
\author{
Charles-André Roux
}
