% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bottom.spectra.R
\name{plot.bottom.spectra}
\alias{plot.bottom.spectra}
\title{This function makes a plot of the upwelling and downwelling
irradiances at the deepest point of the cast. If the bottom
was reached (cops$SHALLOW = TRUE), then a plot of the
calculated bottom reflectance is generated above the initial
graph.}
\usage{
\method{plot}{bottom.spectra}(fullpath, SAVE = F)
}
\arguments{
\item{fullpath}{is the path from the working directory to
the .RData file of the cast}

\item{SAVE}{is a binary parameter specifying whether the
generated plot should be saved. Plots are saved in the
bottom_spectra folder of the working directory. Make sure to
create this directory beforehand.}
}
\description{
This function makes a plot of the upwelling and downwelling
irradiances at the deepest point of the cast. If the bottom
was reached (cops$SHALLOW = TRUE), then a plot of the
calculated bottom reflectance is generated above the initial
graph.
}
\author{
Charles-André Roux
}
