% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cops.R, R/pcops.R
\name{plot.pcops}
\alias{plot.pcops}
\title{S3 plot method for p-cops objects}
\usage{
\method{plot}{pcops}(x, plot.type = c("confplot"), main, asp = NA, ...)

\method{plot}{pcops}(x, plot.type = c("confplot"), main, asp = NA, ...)
}
\arguments{
\item{x}{an object of class cops}

\item{plot.type}{String indicating which type of plot to be produced: "confplot", "reachplot", "resplot","transplot","Shepard","stressplot" (see details)}

\item{main}{the main title of the plot}

\item{asp}{aspect ratio of x/y axis; defaults to NA; setting to 1 will lead to an accurate represenation of the fitted distances.}

\item{...}{Further plot arguments passed: see 'plot.smacof' and 'plot' for detailed information.

Details:
See plot.smacofP}
}
\description{
S3 plot method for p-cops objects

S3 plot method for p-cops objects
}
\examples{
dis<-as.matrix(smacof::kinshipdelta)
resl<-pcops(dis,loss="strain",lower=0.1,upper=5,minpts=2)
plot(resl)
plot(resl,plot.type="Shepard")
dis<-as.matrix(smacof::kinshipdelta)
resl<-pcops(dis,loss="strain",lower=0.1,upper=5,minpts=2)
plot(resl)
plot(resl,plot.type="Shepard")
}
