% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{plot_pvalues}
\alias{plot_pvalues}
\title{Plot p-values in -log10 scale with original labels}
\usage{
plot_pvalues(
  x,
  target,
  x_axis_var = NULL,
  color_var = NULL,
  group_var = NULL,
  palette = "Dark2",
  by = c("Approach", "Embedding", "Clustering", "k"),
  facetx = NULL,
  facety = NULL,
  limits = NULL
)
}
\arguments{
\item{x}{\code{data.frame}}

\item{target}{column name in \code{x} for y axis (boxplot)}

\item{x_axis_var}{column name in \code{x} for x axis}

\item{color_var}{column name in \code{x} for color}

\item{group_var}{column name in \code{x} for group}

\item{palette}{RColorBrewer palette name}

\item{by}{column names in \code{x} to calculate \code{target} quantiles over}

\item{facetx}{column name in \code{x} for x facet}

\item{facety}{column name in \code{x} for y facet}

\item{limits}{y-axis limits}
}
\value{
\code{ggplot} object
}
\description{
To help manage the scale differences, outliers (log p < Q1 - 1.5*IQR) are omitted.
}
