% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{plot_similarity_matrix}
\alias{plot_similarity_matrix}
\title{Plot similarity matrix as a heatmap}
\usage{
plot_similarity_matrix(
  sim_mat,
  method = "average",
  palette = "RdBu",
  limits = c(-1, 1),
  palette_direction = 1,
  title = NULL
)
}
\arguments{
\item{sim_mat}{similarity matrix}

\item{method}{hclust method}

\item{palette}{color distiller palette}

\item{limits}{bounds of similarity}

\item{palette_direction}{set color direction}

\item{title}{plot title}
}
\value{

}
\description{
Order using hierarchical clustering
}
