% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/possol.R
\name{possol}
\alias{possol}
\title{Solar position as a function of date and time UTC and position (lat, lon)}
\usage{
possol(month, day, tu, xlon, xlat)
}
\arguments{
\item{month}{is the month (1 to 12)}

\item{day}{is the day of the month (1-31)}

\item{tu}{is the time UTC in decimal format (0.0 to 23.999)}

\item{xlon}{is the longitude in decimal degrees}

\item{xlat}{is the latitude in decimal degrees}
}
\value{
Returns a vector os two numeric for the zenithal and azimuthal angles in degrees
day is the number of the day in the month
}
\description{
Solar position as a function of date and time UTC and position (lat, lon)
}
\author{
Bernard Gentilly
}
