% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.cops.R
\name{process.cops}
\alias{process.cops}
\title{This the main function of the Cops package and is
called by \code{\link{cops.go}}}
\usage{
process.cops(dirdat, ASCII = FALSE, CLEAN.FILES = FALSE)
}
\arguments{
\item{dirdat}{is the working directory string}

\item{ASCII}{is a logical value: if TRUE the outputs are written in simple ASCII files in a ddition to
the binary files in RData format.}

\item{CLEAN.FILES}{is a logical value: if TRUE, the user will be prompt
to select the good part of the COPS file interactively.
IMPORTANT: the input cops file will be OVERWRITE.}
}
\description{
This the main function of the Cops package and is
called by \code{\link{cops.go}}
}
\author{
Bernard Gentili and Simon Belanger
}
