% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pso-data.R
\docType{data}
\name{psoriasis_clinical_example}
\alias{psoriasis_clinical_example}
\title{Psoriasis clinical data}
\format{
An object of class \code{list}
}
\usage{
data(psoriasis_clinical_example)
}
\description{
Study, patient id, anatomical site when available
}
\keyword{datasets}
