% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pso-data.R
\docType{data}
\name{psoriasis_degs_example}
\alias{psoriasis_degs_example}
\title{Psoriasis example genes}
\format{
An object of class \code{character}
}
\usage{
data(psoriasis_degs_example)
}
\description{
ENSEMBL gene ids
}
\keyword{datasets}
