% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pso-data.R
\docType{data}
\name{psoriasis_example_MEs}
\alias{psoriasis_example_MEs}
\title{Meta-gene features for psoriasis expression data}
\format{
An object of class \code{matrix}
}
\usage{
data(psoriasis_example_MEs)
}
\description{
WGCNA module eigen-genes for psoriasis RNA-Seq data
}
\keyword{datasets}
