% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pso-data.R
\docType{data}
\name{psoriasis_rnaseq_corrected}
\alias{psoriasis_rnaseq_corrected}
\title{Batch corrected psoriasis expression data}
\format{
An object of class \code{matrix}
}
\usage{
data(psoriasis_rnaseq_corrected)
}
\description{
RNA-Seq batch-corrected and normalized counts
}
\keyword{datasets}
