% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pso-data.R
\docType{data}
\name{psoriasis_rnaseq_example}
\alias{psoriasis_rnaseq_example}
\title{Psoriasis expression data}
\format{
An object of class \code{list}
}
\usage{
data(psoriasis_rnaseq_example)
}
\description{
RNA-Seq raw counts
}
\keyword{datasets}
