% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pso-data.R
\docType{data}
\name{psoriasis_rnaseq_normalized}
\alias{psoriasis_rnaseq_normalized}
\title{Normalized psoriasis expression data}
\format{
An object of class \code{matrix}
}
\usage{
data(psoriasis_rnaseq_normalized)
}
\description{
RNA-Seq normalized counts
}
\keyword{datasets}
