% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QWIP.R
\name{QWIP}
\alias{QWIP}
\title{QWIP: A Quantitative Metric for Quality Control of Aquatic Reflectance
Spectral Shape Using the Apparent Visible Wavelength (Dierssen et al., FRS, 2022)}
\usage{
QWIP(Waves, Rrs, PLOT = TRUE, LABEL = "Station")
}
\arguments{
\item{Rrs}{}
}
\value{
QWIP.score
}
\description{
This function calculates the Quality Water Index Polynomial (QWIP)
based on Diesrssen et al 2022 paper.
}
\author{
Simon Bélanger
}
