% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{rbind_fill}
\alias{rbind_fill}
\title{Rbind modification which fills missing columns with NA using base R functions}
\usage{
rbind_fill(a, b)
}
\arguments{
\item{a}{\code{matrix}}

\item{b}{\code{matrix}}
}
\value{
\code{matrix} with mis-matched columns as NA
}
\description{
Rbind modification which fills missing columns with NA using base R functions
}
