% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{reorder_method_factors}
\alias{reorder_method_factors}
\title{Reorder factors in scores to organize plots}
\usage{
reorder_method_factors(x)
}
\arguments{
\item{x}{scores from COPS pipeline}
}
\description{
Reorder factors in scores to organize plots
}
