% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromGenesTo Pathways.R
\name{retrieveDiseaseGenesOT}
\alias{retrieveDiseaseGenesOT}
\title{Retrieve disease-gene associations from the Open Targets platforms.}
\usage{
retrieveDiseaseGenesOT(diseases, fields)
}
\arguments{
\item{diseases}{a character vector indicating the disease names.}

\item{fields}{a character vector indicating the data types used for the infrence of disease-gene associations.
Check the Open Target platforms for more details.}
}
\value{
a data frame including disease-gene association found for each specified data type.
}
\description{
Utility function to extract relevant disease-gene associations
}
