% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromGenesTo Pathways.R
\name{rwr_wrapper}
\alias{rwr_wrapper}
\title{Random walk with restart}
\usage{
rwr_wrapper(
  expr,
  gene_network,
  seed_size = nrow(x)/3,
  absolute_dysregulation = FALSE,
  parallel = 1,
  rwr_restart_probability = 0.75,
  rwr_adjacency_normalization = "laplacian",
  rwr_affinity_normalization = "none",
  ...
)
}
\arguments{
\item{expr}{}

\item{gene_network}{}

\item{seed_size}{}

\item{absolute_dysregulation}{}

\item{parallel}{}

\item{rwr_restart_probability}{}

\item{rwr_adjacency_normalization}{}

\item{rwr_affinity_normalization}{}
}
\value{

}
\description{
Converts gene expression values to network affinity values
}
