% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cops.R
\name{scale_adjust}
\alias{scale_adjust}
\title{Adjusts a configuration}
\usage{
scale_adjust(conf, ref, scale = c("sd", "std", "proc", "none"))
}
\arguments{
\item{conf}{a configuration}

\item{ref}{a reference configuration (only for scale="proc")}

\item{scale}{Scale adjustment. "std" standardizes each column of the configurations to mean=0 and sd=1, "sd" scales the configuration by the maximum standard devation of any column, "proc" adjusts the fitted configuration to the reference}
}
\value{
The scale adjusted configuration.
}
\description{
Adjusts a configuration
}
