% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromGenesTo Pathways.R
\name{seed_similarity}
\alias{seed_similarity}
\title{Visualize gene seed similarity for given settings}
\usage{
seed_similarity(
  dat,
  disease_targets,
  n_top_expr,
  n_top_target,
  mart_results,
  gene.network,
  plot_title,
  otp_score = "association_score.datatypes.rna_expression",
  kernelized_sorting = FALSE,
  reg_up_down = FALSE,
  parallel = 1
)
}
\arguments{
\item{dat}{}

\item{disease_targets}{}

\item{n_top_expr}{}

\item{n_top_target}{}

\item{mart_results}{}

\item{gene.network}{}

\item{plot_title}{}

\item{otp_score}{}

\item{kernelized_sorting}{}

\item{reg_up_down}{}

\item{parallel}{}
}
\value{

}
\description{
Visualize gene seed similarity for given settings
}
