% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwrfgsea.R
\name{seeded_rwr}
\alias{seeded_rwr}
\title{Random walk with restart}
\usage{
seeded_rwr(
  X,
  G,
  p = 0.7,
  graph_normalization = c("laplacian", "transition", "none"),
  cholesky = !igraph::is_directed(G),
  affinity_normalization = TRUE
)
}
\arguments{
\item{X}{a boolean matrix indicating seeds with samples on columns}

\item{G}{an \code{igraph.object} with nodes matching to rows of \code{X}}

\item{p}{restart probability.}

\item{graph_normalization}{method used to normalise the adjacency matrix of the input graph.}

\item{cholesky}{Whether to apply Cholesky decomposition to potentially speed computations. Applicable to undirected graphs.}

\item{affinity_normalization}{controls whether the output matrix is normalised.}
}
\value{
matrix of enrichment scores
}
\description{
Runs random walk in a given network starting from seed genes.
}
