% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{stability_evaluation}
\alias{stability_evaluation}
\alias{stability_evaluation_pac}
\title{Clustering stability evaluation}
\usage{
stability_evaluation(
  clusters,
  by = c("datname", "drname", "run", "k", "m"),
  parallel = 1,
  reference_fold = NULL,
  ...
)

stability_evaluation_pac(
  clusters,
  by = c("datname", "drname", "k", "m"),
  parallel = 1,
  reference_fold = NULL,
  ...
)
}
\arguments{
\item{clusters}{clustering \code{data.frame} such as returned by \code{\link{subsample_clusteval}}}

\item{by}{vector of column names to keep}

\item{parallel}{number of threads}

\item{reference_fold}{fold number that corresponds to reference which other folds are compared against, inferred from input by default}

\item{...}{extra arguments are ignored}
}
\value{
Returns a \code{data.frame} where each row corresponds to clustering stability
        with respect to kept column variables
}
\description{
Performs stability analysis on cross-validated clusterings.
}
\details{
Default settings work with \code{\link{subsample_clusteval}} output 'clusters'.
}
\section{Functions}{
\itemize{
\item \code{stability_evaluation_pac()}: Stability evaluation with Proportion of Ambiguously 
Clustered pairs (PAC)

}}
