% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{subsample_dimred}
\alias{subsample_dimred}
\title{Dimensionality reduction on subsampled data sets}
\usage{
subsample_dimred(
  dat_list,
  sub_index,
  sub_split_data = TRUE,
  parallel = 1,
  by = c("run", "fold"),
  ...
)
}
\arguments{
\item{dat_list}{A list of data.frames or data.tables.}

\item{sub_index}{A data.frame indicating cv folds and runs such as returned by \code{\link{subsampling}}.}

\item{sub_split_data}{Can be set to FALSE if \code{dat_list} elements already contain the columns \code{"run"} and \code{"fold"}.}

\item{parallel}{number of parallel threads}

\item{by}{column names to split data by such that each split corresponds to one dataset and fold}

\item{...}{Extra arguments are passed to \code{\link{dimensionality_reduction}}.}
}
\value{
list of data sets
}
\description{
Dimensionality reduction on subsampled data sets
}
