% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment.R
\name{subsample_pathway_enrichment}
\alias{subsample_pathway_enrichment}
\title{Cross-validation wrapped pathway enrichment}
\usage{
subsample_pathway_enrichment(
  dat_list,
  sub_index,
  gene_id_list,
  parallel = 1,
  ...
)
}
\arguments{
\item{dat_list}{list of data.tables corresponding gene expression matrices where the gene names have been replaced by dimension IDs (dim1, dim2, ...).}

\item{sub_index}{list of data.frames corresponding to cross-validation fold indicators as produced by \code{\link{subsampling}}}

\item{gene_id_list}{list of gene name vectors of the corresponding columns in dat_list}

\item{parallel}{ignored and set to 1 for spawned subprocesses}

\item{...}{arguments passed on to \code{\link[COPS]{genes_to_pathways}}}
}
\value{
\code{list} of \code{data.frame}s with extracted pathway features
}
\description{
This function wraps the pathway enrichment based data transformations for use in a 'cross-validated' clustering pipeline. 
Assumes a parallel backend has been registered for \code{\link[foreach]{foreach}} with \code{\link[doParallel]{registerDoParallel}}. 
Setting \code{parallel} does nothing.
}
