% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_analysis.R
\name{survival_evaluation}
\alias{survival_evaluation}
\title{Likelihood ratio analysis of Cox PH survival models for clustering results}
\usage{
survival_evaluation(
  event_data,
  clusters,
  survival_time_col = "time",
  survival_event_col = "event",
  survival_covariate_names = c("age", "stage"),
  patient_id = "patient",
  ...
)
}
\arguments{
\item{event_data}{}

\item{clusters}{}

\item{survival_time_col}{}

\item{survival_event_col}{}

\item{survival_covariate_names}{}

\item{patient_id}{}

\item{...}{}
}
\value{

}
\description{
Likelihood ratio analysis of Cox PH survival models for clustering results
}
