% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_analysis.R
\name{survival_preprocess}
\alias{survival_preprocess}
\title{Pre-process event data for survival analysis}
\usage{
survival_preprocess(
  event_data,
  event_time_name = "days_to_death",
  follow_up_time_name = "days_to_last_followup",
  event_field_name = "vital_status",
  event_name = "Dead",
  event_time_cutoff = Inf,
  event_time_lower_cutoff = 0
)
}
\arguments{
\item{event_data}{A data.frame that contains survival times, event data and covariates.}

\item{event_time_name}{Name of the column that contains time to event.}

\item{follow_up_time_name}{Name of the column that contains time to end of followup for a censored data point.}

\item{event_field_name}{Name of the column that contains event indicators.}

\item{event_name}{Event column value that corresponds to the event of interest.}

\item{event_time_cutoff}{Upper cutoff value for time.}

\item{event_time_lower_cutoff}{Lower cutoff value for time.}
}
\description{
Use when survival data is not in time and event format e.g., TCGA clinical data.
}
