% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{tsne_viz}
\alias{tsne_viz}
\title{Data visualization using t-SNE}
\usage{
tsne_viz(data, category, category_label, tsne_perplexity = 45)
}
\arguments{
\item{data}{}

\item{category}{}

\item{category_label}{}

\item{tsne_perplexity}{}
}
\value{

}
\description{
Data visualization using t-SNE
}
