% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{umap_viz}
\alias{umap_viz}
\title{Data visualization using UMAP}
\usage{
umap_viz(data, category, category_label, umap_neighbors = 20)
}
\arguments{
\item{data}{}

\item{category}{}

\item{category_label}{}

\item{umap_neighbors}{}
}
\value{

}
\description{
Data visualization using UMAP
}
