% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterization.R
\name{univariate_group_diff_tests}
\alias{univariate_group_diff_tests}
\title{Univariate group difference tests}
\usage{
univariate_group_diff_tests(dat, group, remove_zero_var = TRUE, parallel = 1)
}
\arguments{
\item{dat}{molecular data, samples on columns, rows must be named}

\item{group}{categorical variable matching \code{dat} columns}

\item{remove_zero_var}{whether to remove genes based on \code{\link[caret]{nearZeroVar}}}

\item{parallel}{number of threads}
}
\value{

}
\description{
Computes various statistics for molecular differences between groups.
}
