% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{unsigned_TOM_matrix}
\alias{unsigned_TOM_matrix}
\title{Topology overlap measure}
\usage{
unsigned_TOM_matrix(A)
}
\arguments{
\item{A}{Adjacency \code{matrix}.}
}
\description{
Computes the unsigned topology overlap measure used in WGCNA.
}
