% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_pipeline.R
\name{vertical_pipeline}
\alias{vertical_pipeline}
\title{COPS pipeline vertical parallelization}
\usage{
vertical_pipeline(
  dat_list,
  nfolds = 5,
  nruns = 1,
  survival_data = NULL,
  association_data = NULL,
  multi_omic_methods = NULL,
  parallel = 1,
  data_is_kernels = FALSE,
  silhouette_dissimilarities = NULL,
  ...
)
}
\arguments{
\item{dat_list}{}

\item{nfolds}{}

\item{nruns}{}

\item{survival_data}{}

\item{association_data}{}

\item{multi_omic_methods}{}

\item{parallel}{}

\item{...}{}
}
\value{

}
\description{
COPS pipeline vertical parallelization
}
