% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{weighted_linear_kernel}
\alias{weighted_linear_kernel}
\title{Weighted linear kernel}
\usage{
weighted_linear_kernel(x, weights)
}
\arguments{
\item{x}{feature matrix}

\item{weights}{feature weights}
}
\value{
kernel \code{matrix}
}
\description{
Weighted linear kernel
}
