% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{cbind_fill}
\alias{cbind_fill}
\title{Cbind modification which fills missing rows with NA using base R functions}
\usage{
cbind_fill(a, b)
}
\arguments{
\item{a}{}

\item{b}{}
}
\value{

}
\description{
Cbind modification which fills missing rows with NA using base R functions
}
