% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_cls_pipeline.R
\name{clusteval_scoring}
\alias{clusteval_scoring}
\title{Scoring of dimensionality reduction and clustering pipeline output}
\usage{
clusteval_scoring(
  input,
  by = c("datname", "drname", "k", "m"),
  chisq_significance_level = 0.05,
  w_connectivity = -0.1,
  w_dunn = 0.1,
  w_silhouette = 0.5,
  w_stability = 1,
  w_batch_effect = -1
)
}
\arguments{
\item{input}{\code{\link{dimred_clusteval_pipeline}} output}

\item{by}{character vector containing column names to group analysis by}

\item{chisq_significance_level}{p-value cutoff for computing rejection rate of
\code{chisq.test}}

\item{w_connectivity}{weight for \code{\link[clValid]{connectivity}} score}

\item{w_dunn}{weight for \code{\link[clValid]{dunn}} score}

\item{w_silhouette}{weight for \code{silhouette} score}

\item{w_stability}{weight for \code{\link{stability_eval}} score}

\item{w_batch_effect}{weight for \code{chisq.test} rejection rate,
ignored unless batch_label was supplied as vector or
contained column named "batch_label"}
}
\value{
Returns a \code{list} containing a \code{data.frame} \code{$all} of all scores and
        a single row \code{$best} with the best score
}
\description{
Computes averages of metrics given in pipeline output and also returns the
best combination based on a weighted sum of metrics.
}
\details{
Default weights have
been chosen to emphasize stability and  presence of batch effect. The
internal metrics have wildly varying scales and thus the weights will
need to be guessed manually almost always.
}
