% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromGenesTo Pathways.R
\name{coexpression_network_unweighted}
\alias{coexpression_network_unweighted}
\title{Unweighted gene co-expression network constructor}
\usage{
coexpression_network_unweighted(
  dat,
  correlation_method = "spearman",
  cor_threshold = 0.5
)
}
\arguments{
\item{dat}{}

\item{correlation_method}{}

\item{cor_threshold}{}
}
\value{

}
\description{
Thresholded correlation. Uses \code{WGCNA::signumAdjacencyFunction}.
}
