% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.time.12h.to.24h.string.R
\name{convert.time.12h.to.24h.string}
\alias{convert.time.12h.to.24h.string}
\title{Script to convert the COPS time in UTC in 24h format file}
\usage{
convert.time.12h.to.24h.string(DateTime, cops.init, UTClag)
}
\description{
Script to convert the COPS time in UTC in 24h format file
}
\author{
by Simon Belanger
}
