% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cops.go.R
\name{cops.go}
\alias{cops.go}
\title{COPS routine launcher}
\usage{
cops.go(interactive = FALSE, ASCII = FALSE, CLEAN.FILES = FALSE)
}
\arguments{
\item{interactive}{is a logical value: if TRUE, the user is prompted to
selected a folder containing COPS file to process. The plots are produced in pop-up windows.
If FALSE, you have to create an ASCII file named directories.for.cops.dat and put
the full path of the folders you want to process (on by line). The plots are put in a PDF file
in a sub-folder ./PDF.}

\item{ASCII}{is a logical value: if TRUE the outputs are written in simple ASCII files in a ddition to
the binary files in RData format.}

\item{CLEAN.FILES}{is a logical value: if TRUE, the user will be prompt
to select the good part of the COPS file interactively.
IMPORTANT: the input cops file will be OVERWRITE.}
}
\description{
This the function that should be call to run the whole COPS processing.
}
\author{
Bernard Gentili and Simon Belanger
}
