% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cops-package.R, R/cops.R
\docType{package}
\name{cops}
\alias{cops}
\alias{cops-package}
\title{cops: cluster optimized proximity scaling}
\usage{
cops(dis, variant = c("1", "2", "Variant1", "Variant2", "v1", "v2",
  "COPS-C", "P-COPS", "configuration-c", "profile", "copstress-c",
  "p-copstress", "COPS-P", "copstress-p", "cops-c", "p-cops"), ...)
}
\arguments{
\item{dis}{a dissimilarity matrix or a dist object}

\item{variant}{a character string specifying which variant of COPS to fit. Allowed is any of the following "1","2","Variant1","Variant2","v1","v2","COPS-C","P-COPS","configuration-c","profile","copstress-c","p-copstress". Defaults to "COPS-C".}

\item{...}{arguments to be passed to  \code{\link{copstressMin}} (for Variant 1) or \code{\link{pcops}} (for Variant 2).}
}
\value{
For COPS-C Variant 1 see \code{\link{copstressMin}}, for P-COPS Variant 2 see \code{\link{pcops}}
}
\description{
Cluster optimized proximity scaling (COPS) refers to multidimensional scaling methods that aim at pronouncing the clustered appearance of the configuration. They achieve this by transforming proximities/distances with power functions and augment the fitting criterion with a clusteredness index, the OPTICS Cordillera (Rusch, Hornik & Mair 2017). There are two variants: One for finding the configuration directly for given parameters (COPS-C), and one for using the augmented fitting criterion to find optimal parameters for the power transformations (P-COPS). The package contains various functions, wrappers, methods and classes for fitting, plotting and displaying different MDS models in a COPS framework like Torgerson scaling, SMACOF, Sammon mapping, elastic scaling, symmetric SMACOF, spherical SMACOF, sstress, rstress, powermds, power elastic scaling, power sammon mapping, powerstress. All of these models can also solely be fit as MDS with power transformations. The package further contains functions for optimization (Adaptive LJ Algorithmus).

Minimizing copstress for a clustered MDS configuration. Allows to choose COPS-C (finding a configuration from copstress with cordillera penalty) and profile COPS (finding hyperparameters for MDS models with power transformations). It is wrapper for copstressMin and pcops.
}
\details{
The cops package provides five categories of important functions:

Models & Algorithms:
\itemize{
\item cops() ... high level interface to fit COPS models as described in Rusch et al. (2015). By setting cordweight to zero they can also be used to fit metric MDS for many different models, see below.
\item copstressMin()... The workhorse for fitting a COPS-C model. Can also be called directly.
\item pcops()... The workhorse for fitting a P-COPS model. Can also be called directly.
\item powerStressMin()... a workhorse for fitting s-stress, r-stress (de Leeuw, 2014), p-stress (e.g., Rusch et al., 2015), Sammon mapping with power transformations (powersammon) and elastic scaling with power transformation (powerelastic). They can conveniently also be fitted via the cops functions and setting stressweight=1 and cordweight or by the dedicated functions starting with cops_XXX where XXX is the method and setting stressweight=1 and cordweight=0. It uses the nested majorization algorithm for r-stress of De Leeuw (2014).
}

Optimization functions:
\itemize{
\item ljoptim() ... An (adaptive) version of the Luus-Jakola random search
}

Wrappers and convenience functions:
\itemize{
\item conf_adjust(): procrustes adjustment of configurations 
\item cmdscale(), sammon(): wrappers for that return S3 objects to be used with cops
\item copstress() ... a function to calculate copstress (Rusch et al., 2015)
\item cop_smacofSym(), cop_sammon(), cop_cmdscale(), cop_rstress(), cop_powerstress(), cop_smacofSphere(), cop_sammon2(), cop_elastic(), cop_sstress(), cop_powerelastic(), cop_powersammon(): cop versions of these MDS models.
}

Methods: 
For most of the objects returned by the high-level functions S3 classes and methods for standard generics were implemented, including print, summary, plot, plot3d, plot3dstatic.

References:
\itemize{
\item Rusch, T., Mair, P. \& Hornik, K. (2015) COPS: Cluster optimized proximity scaling, Report 2015/1, Discussion Paper Series, Center for Empirical Research Methods, WU Vienna University of Economics and Business.
\item Rusch, T., Mair, P. \& Hornik, K. (2017) Assessing and quantifying clusteredness: The OPTICS Cordillera. Journal of Computational and Graphical Statistics. Forthcoming. doi = 10.1080/10618600.2017.1349664, \url{http://dx.doi.org/10.1080/10618600.2017.1349664} 
}

Authors: Thomas Rusch, Jan de Leeuw, Patrick Mair

Maintainer: Thomas Rusch
}
\examples{
library(cordillera)
data(BankingCrisesDistances)

\donttest{
#shorthand function for COPS-C (finding configuration with copstress)
res<-cops(BankingCrisesDistances[,1:69],variant="COPS-C",stressweight=0.98,cordweight=0.02)
res
summary(res)
plot(res)
plot(res,"reachplot")
plot(res,"transplot")
plot(res,"Shepard")
#shorthand function for P-COPS (hyperparameter search for powerstress)
res<-cops(BankingCrisesDistances[,1:69],variant="P-COPS")
res
summary(res)
plot(res)
plot(res,"reachplot")
plot(res,"transplot")
plot(res,"Shepard")
}

dis<-as.matrix(smacof::kinshipdelta)

#COPS-C with equal weight to stress and cordillera 
res1<-cops(dis,variant="COPS-C",stressweight=0.5,cordweight=0.5,
minpts=2,itmax=1000) #use higher itmax in real
res1
summary(res1)
plot(res1)
plot(res1,"reachplot")


\donttest{
#s-stress type copstress (i.e. kappa=2, lambda=2)
res3<-cops(dis,variant="COPS-C",kappa=2,lambda=2,stressweight=0.5,cordweight=0.5) 
res3
summary(res3)
plot(res3)


#power-stress type profile copstress
# search for optimal kappa and lambda between kappa=0.5,lambda=0.5 and kappa=2,lambda=5
# nu is fixed on -1
ws<-1/dis
diag(ws)<-1 
res5<-cops(dis,variant="P-COPS",loss="powerstress",
theta=c(1.4,3,-1),lower=c(1,0.5,-1),upper=c(3,5,-1),weightmat=ws,
stressweight=0.9,cordweight=0.1) 
res5
summary(res5)
plot(res5)
}

}
\keyword{clustering}
\keyword{multivariate}
