% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_pipeline.R
\name{COPS}
\alias{COPS}
\title{Clustering algorithms for Omics based Patient Stratification}
\usage{
COPS(
  dat,
  nfolds,
  nruns,
  association_data = NULL,
  survival_data = NULL,
  module_eigs = NULL,
  verbose = TRUE,
  parallel = 1,
  pathway_enrichment_method = "none",
  vertical_parallelization = FALSE,
  ...
)
}
\arguments{
\item{dat}{}

\item{nfolds}{}

\item{nruns}{}

\item{association_data}{}

\item{survival_data}{}

\item{module_eigs}{}

\item{verbose}{}

\item{parallel}{}

\item{pathway_enrichment_method}{}

\item{vertical_parallelization}{}

\item{...}{}
}
\value{

}
\description{
Clustering algorithms for Omics based Patient Stratification
}
