% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cops.R
\name{copstressMin}
\alias{copstressMin}
\title{Fitting a COPS-C Model (COPS Variant 1).}
\usage{
copstressMin(delta, kappa = 1, lambda = 1, nu = 1, theta = c(kappa,
  lambda, nu), weightmat = 1 - diag(nrow(delta)), ndim = 2, init = NULL,
  stressweight = 0.975, cordweight = 0.025, q = 2, minpts = ndim + 1,
  epsilon = 10, dmax = NULL, rang, optimmethod = c("Nelder-Mead",
  "Newuoa"), verbose = 0, scale = c("sd", "rmsq", "std", "proc", "none"),
  normed = TRUE, accuracy = 1e-07, itmax = 1e+05,
  stresstype = c("stress-1", "stress"), ...)
}
\arguments{
\item{delta}{numeric matrix or dist object of a matrix of proximities}

\item{kappa}{power transformation for fitted distances}

\item{lambda}{power transformation for proximities}

\item{nu}{power transformation for weights}

\item{theta}{the theta vector of powers; the first is kappa (for the fitted distances if it exists), the second lambda (for the observed proximities if it exist), the third is nu (for the weights if it exists) . If less than three elements are is given as argument, it will be recycled. Defaults to 1 1 1. Will override any kappa, lmabda, nu parameters if they are given and do not match}

\item{weightmat}{(optional) a matrix of nonnegative weights; defaults to 1 for all off diagonals}

\item{ndim}{number of dimensions of the target space}

\item{init}{(optional) initial configuration}

\item{stressweight}{weight to be used for the fit measure; defaults to 0.975}

\item{cordweight}{weight to be used for the cordillera; defaults to 0.025}

\item{q}{the norm of the corrdillera; defaults to 1}

\item{minpts}{the minimum points to make up a cluster in OPTICS; defaults to ndim+1}

\item{epsilon}{the epsilon parameter of OPTICS, the neighbourhood that is checked; defaults to 10}

\item{dmax}{The winsorization limit of reachability distances in the OPTICS Cordillera. If supplied, it should be either max(rang) or NULL, at which point it is found as 1.5 times the values of the power stress model with the same parameters as the COPS model. If dmax and rang are supplied and dmax is not max(rang), a warning is given and rang takes precedence.}

\item{rang}{range of the reachabilities to be considered. If missing it is found from the initial configuration by taking 0 as the lower boundary and 1.5 times the maximal reachability of the initial fit as upper boundary. If NULL it will be normed to each configuration's minimum and maximum distance as lower and upper bound respectively, so an absolute value of goodness-of-clusteredness. Note that the latter is not necessarily desirable when comparing configurations for their relative clusteredness. See also \code{\link{cordillera}}}

\item{optimmethod}{What optimizer to use? Defaults to NEWUOA, Nelder-Mead is also supported.}

\item{verbose}{numeric value hat prints information on the fitting process; >2 is very verbose}

\item{scale}{Allows to scale the configuration for the OC. One of none (so no scaling), sd (configuration divided by the maximum standard deviation of the columns), std (standardize all columns !NOTE: This does not preserve the relative distances of the optimal config), proc (procrustes adjustment to the initial fit) and rmsq (configuration divided by the maximum root mean square of the columns). Default is sd.}

\item{normed}{should the cordillera be normed; defaults to TRUE}

\item{accuracy}{numerical accuracy, defaults to 1e-8}

\item{itmax}{maximum number of iterations. Defaults to 100000}

\item{stresstype}{which stress to use in the copstress. Defaults to stress-1. If anything else is set, explicitly normed stress which is (stress-1)^2. Using stress-1 puts more weight on MDS fit.}

\item{...}{additional arguments to be passed to the optimization procedure}
}
\value{
A list with the components
        \itemize{
        \item delta: the original transformed dissimilarities
        \item obsdiss: the explicitly normed transformed dissimilarities (which are approximated by the fit)
        \item confdiss: the fitted distances
        \item conf: the configuration to which the scaling of argument scale was applied
        \item confo: the unscaled but explicitly normed configuration returned from the fitting procedure. Scaling applied to confo gives conf.
        \item par, pars : the theta vector of powers tranformations (kappa,lambda,nu)
        \item niter: number of iterations of the optimizer
        \item stress: the square root of explicitly normalized stress (calculated for confo)
        \item spp: stress per point
        \item ndim: number of dimensions
        \item model: Fitted model name with optimizer
        \item call: the call
        \item nobj: the number of objects
        \item type, loss, losstype: stresstype
        \item stress.m, stress.en: other ways to calculate the stress
        \item deltaorig: the original untransformed dissimilarities  
        \item copstress: the copstress loss value
        \item resmat: the matrix of residuals
        \item weightmat: the matrix of untransformed weights 
        \item OC: the (normed) OPTICS Cordillera object (calculated for scaled conf)
        \item OCv: the (normed) OPTICS Cordillera value alone (calculated for scaled conf)
        \item optim: the object returned from the optimization procedure
        \item stressweight, cordweight: the weights of the stress and OC respectively (v_1 and v_2)
        \item optimmethod: The optimizer
}
}
\description{
Minimizing Copstress to obtain a clustered MDS configuration with given hyperparameters theta.
}
\examples{
dis<-as.matrix(smacof::kinshipdelta)

#Copstress with equal weight to stress and cordillera 
res1<-copstressMin(dis,stressweight=0.5,cordweight=0.5) 
res1
summary(res1)
plot(res1)  #super clustered

}
\keyword{clustering}
\keyword{multivariate}
