% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{cv_dimred}
\alias{cv_dimred}
\title{Dimensionality reduction on cross-validated data sets}
\usage{
cv_dimred(dat_list, cv_index, cv_split_data = TRUE, parallel = 1, ...)
}
\arguments{
\item{dat_list}{A list of data.tables.}

\item{cv_index}{A data.frame indicating cv folds and runs such as returned by \code{\link{cv_fold}}.}

\item{cv_split_data}{Can be set to FALSE if \code{dat_list} elements already contain the columns \code{"run"} and \code{"fold"}.}

\item{...}{Extra arguments are passed to \code{\link{dim_reduction_suite}}.}
}
\value{
list of data sets
}
\description{
Dimensionality reduction on cross-validated data sets
}
