% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{cv_fold}
\alias{cv_fold}
\title{Cross-validation fold permutation}
\usage{
cv_fold(
  dat_list,
  nfolds = 5,
  nruns = 2,
  stratified_cv = FALSE,
  anti_stratified = FALSE,
  cv_stratification_var = NULL,
  extra_fold = TRUE,
  ...
)
}
\arguments{
\item{dat_list}{list of datasets, each either data.table or data.frame 
(samples x features) with an "id" column or expression matrix (genes x samples) 
with named columns}

\item{nfolds}{number of cross-validation folds}

\item{nruns}{number of cross-validation replicates}

\item{stratified_cv}{if \code{TRUE}, perform stratified sampling for folds}

\item{anti_stratified}{if \code{TRUE}, maximize separation of batch labels 
within folds, opposite of stratified sampling}

\item{cv_stratification_var}{labels used for stratification}

\item{extra_fold}{if \code{TRUE}, generates an extra fold (nfolds+1) that 
corresponds to the full dataset}

\item{...}{extra arguments are ignored}
}
\value{
list of data.frames with added columns "fold", "run" and "cv_index" as well as 
        duplicated rows of the original data corresponding to different folds.
}
\description{
Creates cross-validation folds of data for downstream analysis.
}
