% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_pipeline.R
\name{dimred_clusteval_pipeline}
\alias{dimred_clusteval_pipeline}
\title{Pipeline combining dimensionality reduction, clustering, cross-validation and evaluation}
\usage{
dimred_clusteval_pipeline(
  dat,
  nfolds,
  nruns,
  batch_label = NULL,
  subtype_label = NULL,
  clinical_data = NULL,
  survival_data = NULL,
  module_eigs = NULL,
  verbose = TRUE,
  parallel = 1,
  pathway_enrichment_method = "none",
  ...
)
}
\arguments{
\item{dat}{A single matrix or list of data matrices corresponding to the same data but different pre-processing.}

\item{nfolds}{Number of cross-validation folds for stability evaluation and metric estimates.}

\item{nruns}{Number of cross-validation replicates for stability evaluation and metric estimates.}

\item{batch_label}{Vector or matrix with categorical variables on columns.}

\item{subtype_label}{Vector or matrix with categorical variables on columns.}

\item{survival_data}{Data for survival analysis, see \code{\link{survival_preprocess}} for details.}

\item{module_eigs}{Module eigen genes for gene module correlation evaluation (see \code{\link{gene_module_score}} for details).}

\item{verbose}{If \code{TRUE}, prints progress notifications.}

\item{parallel}{Sets up and registers \code{parallel} number of threads for supported operations.}

\item{pathway_enrichment_method}{\code{enrichment_method} for \code{\link{genes_to_pathways}}}

\item{...}{extra arguments are passed to pipeline components where appropriate}
}
\value{
Returns a \code{list} of pipeline component outputs for given settings and input data sets
}
\description{
Combines \code{\link{cv_fold}}, \code{\link{cv_dimred}}, \code{\link{cv_clusteval}} and
\code{\link{stability_eval}}.
}
\examples{
library(parallel)
library(COPS)

# DR-CL
res <- dimred_clusteval_pipeline(ad_ge_micro_zscore, 
batch_label = ad_studies, 
parallel = 2, nruns = 2, nfolds = 5, 
dimred_methods = c("pca", "umap", "tsne"), 
cluster_methods = c("hierarchical", "kmeans"), 
metric = "euclidean", 
n_clusters = 2:4)

# CL
res <- dimred_clusteval_pipeline(ad_ge_micro_zscore, 
batch_label = ad_studies, 
parallel = 2, nruns = 2, nfolds = 5, 
dimred_methods = c("none"), 
cluster_methods = c("hierarchical"), 
metric = "correlation", 
n_clusters = 2:4)

# BK-CL
res <- dimred_clusteval_pipeline(ad_ge_micro_zscore, 
batch_label = ad_studies, 
pathway_enrichment_method = "DiffRank", 
gene_key_expr = "ENSEMBL", 
gs_subcats = "CP:KEGG", 
parallel = 2, nruns = 2, nfolds = 5, 
dimred_methods = c("none"), 
cluster_methods = c("hierarchical"), 
metric = "correlation", 
n_clusters = 2:4)

}
