% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_cls_pipeline.R
\name{dimred_clusteval_pipeline}
\alias{dimred_clusteval_pipeline}
\title{Pipeline combining dimensionality reduction, clustering, cross-validation and evaluation}
\usage{
dimred_clusteval_pipeline(
  dat_list,
  nfolds,
  nruns,
  batch_label = NULL,
  verbose = TRUE,
  parallel = 1,
  ...
)
}
\arguments{
\item{dat_list}{list of pre-processed data sets to run pipeline for}

\item{nfolds}{number of cross-validation folds}

\item{nruns}{number of cross-validation replicates}

\item{batch_label}{vector or matrix with categorical variables on columns}

\item{verbose}{if \code{TRUE}, prints progress notifications}

\item{parallel}{sets up and registers \code{parallel} number of threads for supported operations}

\item{...}{extra arguments are passed to pipeline components where appropriate}
}
\value{
Returns a \code{list} of pipeline component outputs for given settings and input data sets
}
\description{
Combines \code{\link{cv_fold}}, \code{\link{cv_dimred}}, \code{\link{cv_clusteval}} and
\code{\link{stability_eval}}.
}
