% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{expressionToRWFeatures}
\alias{expressionToRWFeatures}
\title{Gene expression to gene seeds to RWR on PPI into FGSEA scores}
\usage{
expressionToRWFeatures(
  dat,
  disease_id,
  otp_score = "association_score.datatypes.rna_expression",
  otp_cutoff = 0,
  ppi_cutoff = 700,
  pw_min.size = 5,
  pw_max.size = 200,
  dat_gene_key = "SYMBOL",
  gs_subcats = c("BP", "MF", "CP:KEGG", "CP:REACTOME"),
  directed_ppi = TRUE,
  ...
)
}
\arguments{
\item{dat}{a gene expression matrix with samples on columns}

\item{disease_id}{integer ID in Open Targets platform}

\item{otp_cutoff}{numeric association score cutoff for Open Targets platform}

\item{ppi_cutoff}{numeric PPI link score cutoff}

\item{...}{extra arguments are passed on to \code{\link{fromGeneToNetworksToPathwayFeatures}}}
}
\value{
list of enrichment scores
}
\description{
Conveniently wraps \code{\link{fromGeneToNetworksToPathwayFeatures}} using 
\code{\link{retrieveDiseaseGenesOT}} for disease associated genes, 
\code{STRINGdb} human PPI as a network, \code{msigdbr} for gene set annotations and 
\code{biomaRt} for harmonization.
}
