% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromGenesTo Pathways.R
\name{fgsea_wrapper}
\alias{fgsea_wrapper}
\title{FGSEA wrapper}
\usage{
fgsea_wrapper(data_matrix, list_db_annots, rwr_cutoff = 0, parallel = 1, ...)
}
\arguments{
\item{data_matrix}{}

\item{list_db_annots}{}

\item{rwr_cutoff}{}

\item{parallel}{}
}
\value{

}
\description{
FGSEA wrapper
}
