% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{gene_module_score}
\alias{gene_module_score}
\title{Gene module score}
\usage{
gene_module_score(clust, module_eigs, threshold = 0.3, nan.substitute = 0)
}
\arguments{
\item{clust}{A \code{data.frame} with columns "id" and "cluster".}

\item{module_eigs}{Gene module eigen-genes for each sample (samples x modules).}

\item{threshold}{Threshold for counting correlations.}

\item{nan.substitute}{Substituted value when dividing by zero when there are no correlated clusters for a module.}
}
\description{
Metric based on gene module eigen gene correlation
}
\examples{
library(COPS)
library(WGCNA)
library(parallel)

# Generate module eigen genes with WGCNA
gene_correlation <- cor(t(ad_ge_micro_zscore), method =  "spearman")
adj <- WGCNA::adjacency.fromSimilarity(gene_correlation, power = 2)
TOM <- WGCNA::TOMsimilarity(adj, TOMType = "unsigned")
geneTree <- flashClust::flashClust(as.dist(1 - TOM), method="average")
dynamicMods <- dynamicTreeCut::cutreeDynamic(dendro = geneTree,  method="tree", minClusterSize = 20)
adj_modules <- WGCNA::adjacency.fromSimilarity(gene_correlation[dynamicMods != 0, dynamicMods != 0], power = 2)
MEList <- WGCNA::moduleEigengenes(t(ad_ge_micro_zscore[dynamicMods != 0,]), colors = dynamicMods[dynamicMods != 0])
MEs <- MEList$eigengenes

# Compute the module score
clust <- cutree(hclust(as.dist(1-cor(ad_ge_micro_zscore, method = "spearman")), method = "average"), k = 3)
clust <- data.frame(id = names(clust), cluster = clust)

score <- gene_module_score(clust, MEs)

# Within full pipeline
res <- dimred_clusteval_pipeline(ad_ge_micro_zscore, 
batch_label = ad_studies, 
parallel = 2, nruns = 2, nfolds = 5, 
dimred_methods = c("pca", "umap", "tsne"), 
cluster_methods = c("hierarchical", "kmeans"), 
metric = "euclidean", 
n_clusters = 2:4, 
module_eigs = MEs)

scores <- clusteval_scoring(res, wsum = Silhouette - Module_score, summarise = TRUE)
}
