% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromGenesTo Pathways.R
\name{genes_to_pathways}
\alias{genes_to_pathways}
\title{Transform a gene-level data matrix into path-level information}
\usage{
genes_to_pathways(
  dat,
  enrichment_method = "GSVA",
  db_annots = NULL,
  batch_label_pw = NULL,
  min.size = 5,
  max.size = 200,
  parallel = 1,
  verbose = FALSE,
  key_name = "SYMBOL",
  kcdf = "Gaussian",
  gs_subcats = c("BP", "MF", "CP:KEGG", "CP:REACTOME"),
  rwr_ecdf = FALSE,
  ...
)
}
\arguments{
\item{dat}{a numeric matrix representing gene expression profiles. Genes on the rows and samples on the columns.}

\item{enrichment_method}{options: DiffRank GSVA, RWRFGSEA}

\item{db_annots}{data.frame of pathway annotations or gene sets (e.g. as returned by \code{\link[msigdbr]{msigdbr}})}

\item{batch_label_pw}{batch labels for batch-wise enrichment, ignored if NULL}

\item{min.size}{a numeric value indicating the minimum size of gene sets included}

\item{max.size}{a numeric value indicating the maximum size of gene sets included}

\item{parallel}{a numeric value indicating the number of processors to use when doing the calculations in parallel.}

\item{verbose}{controls verbosity}

\item{key_name}{column name to look for if gene id translation is needed, ids should match data ids rownames \code{\link[org.Hs.eg.db]{org.Hs.eg.db}}}

\item{kcdf}{distribution name for \code{\link[GSVA]{gsva}} empirical distribution kernel}

\item{gs_subcats}{a character vector indicating \code{\link[msigdbr]{msigdbr}}) gene set subcategory names to include in the analysis}

\item{rwr_ecdf}{controls whether a kernel based empirical density function is used when selecting gene seeds for RWR in \code{\link[COPS]{rwr_wrapper}}}

\item{...}{extra arguments are passed to \code{\link[COPS]{rwr_wrapper}} and \code{\link[COPS]{fgsea_wrapper}}}
}
\value{
a list of data.frames corresponding to the transformed features based on the selected gene sets (only supports GO, KEGG and REACTOME at the moment)
}
\description{
This is a utility function which wraps several methods to extract pathway-based features from gene expression data
}
