% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromGenesTo Pathways.R
\name{genes_to_pathways}
\alias{genes_to_pathways}
\title{Transform a gene-level data matrix into path-level information}
\usage{
genes_to_pathways(
  dat,
  enrichment_method = "GSVA",
  db_annots = NULL,
  batch_label_pw = NULL,
  min.size = 5,
  max.size = 200,
  parallel = 1,
  verbose = FALSE,
  key_name = "SYMBOL",
  kcdf = "Gaussian",
  gs_subcats = c("BP", "MF", "CP:KEGG", "CP:REACTOME"),
  rwr_ecdf = FALSE,
  ...
)
}
\arguments{
\item{dat}{a numeric matrix representing gene expression profiles. Genes on the rows and samples on the columns.}

\item{db_annots}{}

\item{min.size}{a numeric value indicating the minimum size of gene sets included}

\item{max.size}{a numeric value indicating the maximum size of gene sets included}

\item{parallel}{a numeric value indicating the number of processors to use when doing the calculations in parallel.}

\item{verbose}{controls verbosity}

\item{key_name}{for gene id translation, the ids should match data ids rownames \code{\link[org.Hs.eg.db]{org.Hs.eg.db}}}

\item{kcdf}{distribution name for \code{\link[GSVA]{gsva}} empirical distribution kernel}

\item{gs_subcats}{a character vector indicating msigdbr gene set subcategory names to include in the analysis}

\item{...}{}

\item{method}{}

\item{batch_label}{}
}
\value{
a list of three data frames:\cr
       - \strong{KEGG_PW}: a character variable containing entrez gene ids;\cr
       - \strong{GO_PW}: a character variable containing gene symbols;\cr
       - \strong{REACTOME_PW}: a numeric variable containing gene-disease scores.
}
\description{
Utility function to extract pathway-based features from gene expression data using GSVA.
}
