% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_enrichment.R
\name{genes_to_pathways}
\alias{genes_to_pathways}
\alias{DiffRank}
\title{Transform a gene-level data matrix into path-level information}
\usage{
genes_to_pathways(
  expr,
  enrichment_method = "GSVA",
  gene_set_list = NULL,
  batch_label_pw = NULL,
  min_size = 5,
  max_size = 200,
  parallel = 1,
  verbose = FALSE,
  gene_key_expr = "SYMBOL",
  gs_subcats = c("GO:BP", "GO:MF", "CP:KEGG", "CP:REACTOME"),
  gsva_kcdf = "Gaussian",
  ...
)

DiffRank(expr, gene_set_list, parallel = 1)
}
\arguments{
\item{expr}{gene expression matrix, samples on columns and genes on rows}

\item{enrichment_method}{options: "DiffRank", "GSVA", "RWRFGSEA"}

\item{gene_set_list}{list of gene sets with gene names that correspond to rows in \strong{expr}}

\item{batch_label_pw}{batch labels for batch-wise enrichment, normal enrichment performed if NULL}

\item{min_size}{a numeric value indicating the minimum size of gene sets included}

\item{max_size}{a numeric value indicating the maximum size of gene sets included}

\item{parallel}{a numeric value indicating the number of processors to use when doing the calculations in parallel.}

\item{verbose}{controls verbosity}

\item{gene_key_expr}{if \code{is.null(gene_set_list)} and \code{expr} rownames are not gene symbols, this specifies the column name in 
\code{\link[org.Hs.eg.db]{org.Hs.eg.db}} to translate pathway gene symbols to. The default value results in gene symbol based gene 
sets when \code{is.null(gene_set_list)}.}

\item{gs_subcats}{if \code{is.null(gene_set_list)}, this character vector indicates \code{\link[msigdbr]{msigdbr}}) gene set subcategory 
names that are included in the analysis.}

\item{gsva_kcdf}{distribution name for \code{\link[GSVA]{gsva}} empirical probability distribution kernel.}

\item{...}{extra arguments are passed to \code{\link[COPS]{RWRFGSEA}}}
}
\value{
a list of data.frames corresponding to the transformed features based on the selected gene sets (only supports GO, KEGG and REACTOME at the moment)


}
\description{
This is a utility function which wraps several methods to extract pathway-based features from gene expression data
}
\section{Functions}{
\itemize{
\item \code{DiffRank}: DiffRank by Wang et al. BMC Medical Genomics 2019
}}

\examples{
library(parallel)
library(COPS)

## GSVA example
ad_gsva <- genes_to_pathways(as.matrix(ad_ge_micro_zscore), "GSVA", 
                             parallel = 2, gene_key_expr = "ENSEMBL", 
                             gs_subcats = "CP:KEGG")
# batch-wise
ad_gsva <- genes_to_pathways(as.matrix(ad_ge_micro_zscore), "GSVA", 
                             batch_label_pw = ad_studies$GSE, parallel = 2, 
                             gene_key_expr = "ENSEMBL", gs_subcats = "CP:KEGG")

## DiffRank example with batch-wise enrichment
ad_diffrank <- genes_to_pathways(ad_ge_micro_zscore, "DiffRank", 
                                 parallel = 2, gene_key_expr = "ENSEMBL", 
                                 gs_subcats = "CP:KEGG")
# batch-wise
ad_diffrank <- genes_to_pathways(ad_ge_micro_zscore, "DiffRank", 
                                 batch_label_pw = ad_studies$GSE, 
                                 parallel = 2, gene_key_expr = "ENSEMBL", 
                                 gs_subcats = "CP:KEGG")

## RWRFGSEA example
ad_data <- ad_ge_micro_zscore
rownames(ad_data) <- as.character(AnnotationDbi::mapIds(org.Hs.eg.db::org.Hs.eg.db, 
                                                        rownames(ad_data), "SYMBOL", "ENSEMBL"))
ad_wgcna_net <- coexpression_network_unweighted(ad_data)
kegg_annotations <- msigdbr::msigdbr(species = "Homo sapiens", 
                                     category = "C2", 
                                     subcategory = "CP:KEGG")
list_kegg_annotations <- lapply(split(kegg_annotations, kegg_annotations$gs_name), 
                                function(x) x$human_gene_symbol) 

ad_rwrfgsea <- genes_to_pathways(ad_data, "RWRFGSEA", gene_set_list = list_kegg_annotations, 
                                 gene_network = ad_wgcna_net, parallel = 2)

}
