% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{get_multi_omic_kernels}
\alias{get_multi_omic_kernels}
\title{Multi-omic kernels}
\usage{
get_multi_omic_kernels(
  dat_list,
  kernels = rep_len("linear", length(dat_list)),
  kernels_center = TRUE,
  kernels_normalize = TRUE,
  kernels_scale_norm = FALSE,
  kernel_gammas = rep_len(0.5, length(dat_list)),
  pathway_networks = NULL,
  pathway_node_betweenness_endpoints = TRUE,
  pathway_first_shortest_path = FALSE,
  pamogk_restart = 0.7,
  pamogk_seeds = "discrete",
  pamogk_seed_under_threshold = qnorm(0.025),
  pamogk_seed_over_threshold = qnorm(0.975),
  pamogk_rwr_verbose = FALSE,
  gene_id_list = NULL,
  zero_var_removal = TRUE,
  mvc_threads = 1,
  preprocess_data = FALSE,
  experimental_parallelization = FALSE,
  ...
)
}
\arguments{
\item{dat_list}{List of input \code{data.frame}s for input.}

\item{kernels}{Character vector of kernel names to use for different views. 
See details.}

\item{kernels_center}{Logical vector specifying which kernels should be 
centered. Repeated for each view if length 1.}

\item{kernels_normalize}{Logical vector specifying which kernels should be 
normalized Repeated for each view if length 1.}

\item{kernels_scale_norm}{Logical vector specifying which kernels should be 
scaled to unit F-norm. Repeated for each view if length 1.}

\item{kernel_gammas}{Numeric vector specifying gamma for the gaussian kernel.}

\item{pathway_networks}{List of \code{igraph} objects containing pathway 
networks. Required for pathway kernels.}

\item{pathway_node_betweenness_endpoints}{whether to include shortest path 
endpoints in betweenness. Including it results in more non-zero weights 
in BWK and PAMOGK.}

\item{pamogk_restart}{Restart probability for PAMOGK RWR.}

\item{pamogk_seeds}{Seed selection strategy for PAMOGK RWR, one of: 
"discrete", "continuous", "threshold". See details below.}

\item{pamogk_seed_under_threshold}{z-score threshold for under-expressed.}

\item{pamogk_seed_over_threshold}{z-score threshold for over-expressed.}

\item{pamogk_rwr_verbose}{See \code{\link[dnet]{dRWR}}}

\item{zero_var_removal}{If set, removes all zero variance features 
from the data. It is called fold-wise, because this is assumed to be run 
inside CV.}

\item{mvc_threads}{Number of threads to use for supported operations.}

\item{...}{Extra arguments are ignored.}
}
\value{
\code{list} of kernels

Supported kernels: 
\itemize{
  \item "linear" - Linear kernel based on standard dot product. 
  \item "gaussian", "rbf" - Gaussian kernel, a.k.a. radial basis function.
  \item "jaccard" - Kernel based on Jaccard index. Used for binary features. 
  \item "tanimoto" - For now, this is identical to "jaccard".
  \item "BWK" - Betweenness Weighted Kernel. Uses pathway networks to compute 
    betweenness centrality which is used to weight features in a linear 
    pathway kernels. 
  \item "PAMOGK" - PAthway Multi-Omics Graph Kernel (Tepeli et al. 2021). 
    Uses z-scores, RWR and shortest paths in pathway networks to create 
    pathway kernels. 
  \item "PIK" - Pathway Induced Kernel (Manica et al. 2019). Uses pathway 
    network adjacency matrices (specifically normalized Laplacians) to define 
    pathway kernels. 
}
Please note that for pathway kernels, the input data must always be mapped to 
genes and that the names must match with the gene names in the pathways. 
The default set of pathways is KEGG molecular pathways with gene symbols. 

PAMOGK RWR seed weight options:
\itemize{
  \item "discrete" - 1 if |z| > t, 0 otherwise. 
  \item "continuous" - z
  \item "threshold" - z if |z| > t, 0 otherwise
}
Regardless of the option, the seeds are divided into two sets based on the 
sign of the z-score. Each set has a separate smoothing step and the end 
result is two different kernels per pathway per omic. This impacts the 
RWR label smoothing by changing the initial distribution.
}
\description{
Process a set of omics observations into kernels
}
