% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getHumanPPIfromSTRINGdb}
\alias{getHumanPPIfromSTRINGdb}
\title{Retrieve human protein-protein interaction network from STRINGdb.}
\usage{
getHumanPPIfromSTRINGdb(gene.diseases, cutoff = 700, directed = FALSE)
}
\arguments{
\item{gene.diseases}{a character vector indicating the gene seeds.}

\item{cutoff}{a numeric value indicating the cutoff for the edge scores.}

\item{directed}{a boolean value indicating the type of grpah to be generated.}
}
\value{
an igraph object.
}
\description{
Utility function to extract a gene subnetwork from STRINGdb including only the seed genes and their interactions.
}
