% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{global_kernel_kmeans}
\alias{global_kernel_kmeans}
\title{Global kernel k-means (Tzortzis & Likas 2008)}
\usage{
global_kernel_kmeans(K, n_clusters)
}
\arguments{
\item{K}{kernel \code{matrix}}

\item{n_clusters}{number of clusters}
}
\value{
\code{list} of cluster assignments and k-means objective
}
\description{
Global kernel k-means (Tzortzis & Likas 2008)
}
