% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromGenesTo Pathways.R
\name{GSVA}
\alias{GSVA}
\title{Pathway enrichment by using GSVA}
\usage{
GSVA(
  dat,
  list_db_annots = NULL,
  batch_label_pw = NULL,
  min.size = 5,
  max.size = 200,
  study_batch = NULL,
  key_name = "SYMBOL",
  kcdf = "Gaussian",
  parallel = 1,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{dat}{a numeric matrix representing gene expression profiles. Genes on the rows and samples on the columns.}

\item{min.size}{a numeric value indicating the minimum size of gene sets included}

\item{max.size}{a numeric value indicating the maximum size of gene sets included}

\item{study_batch}{factor indicating the individual studies. 
A NULL value implicates that the input data matrix represents the original dataset (without removing the batches).}

\item{kcdf}{distribution name for \code{\link[GSVA]{gsva}} empirical distribution kernel}

\item{parallel}{a numeric value indicating the number of processors to use when doing the calculations in parallel.}

\item{verbose}{controls verbosity}

\item{gs_subcats}{a character vector indicating msigdbr gene set subcategory names to include in the analysis}
}
\value{
a list of data.frames corresponding to the transformed features based on the given gene sets split into KEGG, GO and REACTOME
}
\description{
Utility function wrapping GSVA, batch-wise enrichment and gene key matching.
}
