% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterization.R
\name{heatmap_annotated}
\alias{heatmap_annotated}
\title{Annotated heatmap}
\usage{
heatmap_annotated(
  dat,
  variable_list = list(),
  feature_names = NULL,
  center = TRUE,
  scale = FALSE,
  show_column_names = FALSE,
  show_row_names = TRUE,
  show_column_dend = FALSE,
  show_row_dend = FALSE,
  row_names_side = "left",
  legend_names = NULL,
  color_breaks = NULL,
  col_dat = circlize::colorRamp2(color_breaks, c("blue", "white", "red")),
  ...
)
}
\arguments{
\item{dat}{feature matrix or list of matrices, samples on columns}

\item{variable_list}{list of column annotation variables}

\item{feature_names}{optional character vector matching to a subset of \code{dat} rownames}

\item{center}{whether to apply row-wise center, passed to \link[base]{scale}}

\item{scale}{whether to apply row-wise scaling, passed to \link[base]{scale}}

\item{show_column_names}{whether to include colnames (sample names)}

\item{show_row_names}{whether to include rownames (gene names)}

\item{show_column_dend}{whether to include column dendrogram}

\item{show_row_dend}{whether to include row dendrogram}

\item{row_names_side}{where row names should be located}

\item{legend_names}{names to use for \code{dat} color legends}

\item{color_breaks}{manual color breaks in form \code{c(min, middle, max)}, 
passed to \link[circlize]{colorRamp2}}

\item{...}{other arguments passed to \link[ComplexHeatmap]{Heatmap}}
}
\value{
\code{\link[ComplexHeatmap]{`HeatmapList-class`}}
}
\description{
Generate a heatmap with annotations.
}
