% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{jaccard_similarity}
\alias{jaccard_similarity}
\title{Jaccard similarity coefficient between two partitions}
\usage{
jaccard_similarity(a, b)
}
\arguments{
\item{a}{group indicators corresponding to a data partition}

\item{b}{group indicators corresponding to another data partition}
}
\value{
Jaccard similarity coefficient
}
\description{
Jaccard similarity coefficient between two partitions
}
