% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{kernel_kmeans}
\alias{kernel_kmeans}
\title{Kernel k-means}
\usage{
kernel_kmeans(K, n_k, n_initializations = 100, ...)
}
\arguments{
\item{K}{}

\item{n_k}{}

\item{n_initializations}{}

\item{...}{}
}
\value{

}
\description{
K-means++ with set number of iterations.
}
