% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{mkkm_mr}
\alias{mkkm_mr}
\title{Multiple kernel k-means with matrix induced regularization (Liu et al. 2016)}
\usage{
mkkm_mr(K_list, k, lambda, tolerance = 1e-06, parallel = 0, use_mosek = FALSE)
}
\arguments{
\item{K_list}{}

\item{k}{}

\item{lambda}{}

\item{tolerance}{}
}
\description{
Multiple kernel k-means with matrix induced regularization (Liu et al. 2016)
}
