% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_omics.R
\name{multi_omic_clustering}
\alias{multi_omic_clustering}
\title{Multi-omic clustering via multi-view clustering or integration}
\usage{
multi_omic_clustering(
  dat_list_clust,
  non_data_cols,
  multi_view_methods = "iClusterPlus",
  n_clusters = 2,
  distance_metric = "euclidean",
  correlation_method = "spearman",
  nmf_maxiter = 200,
  nmf_st.count = 20,
  nmf_n.ini = 30,
  nmf_ini.nndsvd = TRUE,
  mofa_scale_views = FALSE,
  mofa_convergence_mode = "medium",
  mofa_maxiter = 1000,
  mofa_environment = NULL,
  anf_neighbors = 20,
  ...
)
}
\arguments{
\item{dat_list_clust}{}

\item{non_data_cols}{}

\item{multi_view_methods}{}

\item{n_clusters}{}
}
\value{

}
\description{
Multi-omic clustering via multi-view clustering or integration
}
