% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_omics.R
\name{multi_omic_clustering}
\alias{multi_omic_clustering}
\title{Multi-omic clustering via multi-view clustering or integration}
\usage{
multi_omic_clustering(
  dat_list_clust,
  non_data_cols,
  multi_view_methods = "iClusterPlus",
  n_clusters = 2,
  distance_metric = "euclidean",
  correlation_method = "spearman",
  standardize_data = FALSE,
  non_negativity_transform = rep_len("none", length(dat_list_clust)),
  icp_view_types = rep_len("gaussian", length(dat_list_clust)),
  icp_bayes_burnin = 1000,
  icp_bayes_draw = 1200,
  nmf_maxiter = 200,
  nmf_st.count = 20,
  nmf_n.ini = 30,
  nmf_ini.nndsvd = TRUE,
  mofa_scale_views = FALSE,
  mofa_likelihoods = rep_len("gaussian", length(dat_list_clust)),
  mofa_convergence_mode = "medium",
  mofa_maxiter = 1000,
  mofa_environment = NULL,
  mofa_lib_path = NULL,
  anf_neighbors = 20,
  kernels = rep_len("linear", length(dat_list_clust)),
  kernels_center = TRUE,
  kernels_normalize = TRUE,
  kernel_gammas = rep_len(0.5, length(dat_list_clust)),
  pathway_networks = NULL,
  pamogk_restart = 0.7,
  kkmeans_maxiter = 100,
  kkmeans_n_init = 100,
  mkkm_mr_lambda = 1,
  mkkm_mr_tolerance = 1e-08,
  mkkm_mr_initialization = TRUE,
  mkkm_mr_mosek = FALSE,
  ecmc_a = 1,
  ecmc_b = 1,
  ecmc_eps = 1e-06,
  ecmc_maxiter = 100,
  ecmc_solver = "MOSEK",
  ecmc_mkkm_mr = TRUE,
  data_is_kernels = FALSE,
  foldwise_zero_var_removal = TRUE,
  mvc_threads = 1,
  gene_id_list = NULL,
  ...
)
}
\arguments{
\item{dat_list_clust}{}

\item{non_data_cols}{}

\item{multi_view_methods}{}

\item{n_clusters}{}
}
\value{

}
\description{
Multi-omic clustering via multi-view clustering or integration
}
