% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pareto_plot}
\alias{pareto_plot}
\title{Pairwise plots for visual Pareto based multi-objective optimization}
\usage{
pareto_plot(
  scores,
  plot_palette = pals::watlington(16),
  metrics = c("TrainStabilityJaccard", "Silhouette", "Smallest_cluster_size"),
  metrics_scale = rep("identity", length(metrics)),
  color_var = "drname",
  shape_var = "m",
  size_var = "k",
  size_range = c(2, 6)
)
}
\arguments{
\item{scores}{}

\item{plot_palette}{color codes used for coloring based on \code{color_var}}

\item{metrics}{}

\item{color_var}{}

\item{shape_var}{}

\item{size_var}{}
}
\value{

}
\description{
Pairwise plots for visual Pareto based multi-objective optimization
}
