% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{pareto_plot}
\alias{pareto_plot}
\title{Pairwise plots for visual Pareto based multi-objective optimization}
\usage{
pareto_plot(
  scores,
  plot_palette = pals::watlington(16),
  metrics = c("TrainStabilityJaccard", "Silhouette", "Smallest_cluster_size"),
  metrics_scale = rep("identity", length(metrics)),
  color_var = Transform,
  shape_var = Clustering,
  size_var = k,
  size_range = c(2, 6),
  color_scale = ggplot2::scale_color_manual(values = plot_palette),
  plot_pareto_front = FALSE,
  front_color = "black",
  metric_comparators = if (plot_pareto_front) get_metric_comparators(metrics) else NULL,
  point_args = list()
)
}
\arguments{
\item{scores}{\code{data.frame} of scores}

\item{plot_palette}{color codes used for coloring based on \code{color_var}}

\item{metrics}{column names in \code{scores} to plot}

\item{metrics_scale}{either "identity" or "nlog10" where the latter is meant 
for translating p-values to negative log10}

\item{color_var}{column name in \code{scores} for color}

\item{shape_var}{column name in \code{scores} for shape}

\item{size_var}{column name in \code{scores} for size}

\item{size_range}{controls the range of point sizes}

\item{color_scale}{can be used instead of \code{plot_palette} to control colors}

\item{plot_pareto_front}{Whether to plot a step-function showing the first Pareto front for each pair of metrics.}

\item{metric_comparators}{Required if \code{plot_pareto_front==TRUE}, see details below.}
}
\value{
\code{\link[gridExtra]{grid.arrange}} result
}
\description{
Plots the given data with pairwise scatter plots for each pair of given columns. 
Intended to be used with \code{\link{scoring}} output.
}
\details{
For Pareto fronts \code{metric_comparators} should be a list of \code{.Primitive(">")} and \code{.Primitive("<")}. 
">" should be used for metrics that should be maximized and 
"<" for metrics that should minimized. By default they are guessed by 
\code{\link{get_metric_comparators}}, but for plots involving variable 
associations and p-values they should be set manually.
}
