% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.Rrs.Kd.for.station.R
\name{plot.Rrs.Kd.for.station}
\alias{plot.Rrs.Kd.for.station}
\title{This function plot the replicates of Rrs and
the integrated Kd (K0) from surface to a given depth
for the cast retained (i.e. value of one in the remove.cops.dat)
It reads the files "remove.cops.dat" and then load the RData file
that are stored in the "./BIN/" directory.}
\usage{
\method{plot}{Rrs.Kd.for.station}(path = "./", depthEdZ = 1)
}
\description{
@param path is the path where raw COPS data are located.
@param depthEdZ is the depth of the ingetration of Kd. Default is 1 meter
}
\details{
@return It produces 2 plots in PNG format in the working directory.
}
\author{
Simon Belanger, UQAR
}
