% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{plot_pvalues}
\alias{plot_pvalues}
\title{Plot p-values in -log10 scale with original labels}
\usage{
plot_pvalues(
  x,
  target,
  x_axis_var = NULL,
  color_var = NULL,
  group_var = NULL,
  palette = "Dark2",
  by = c("Approach", "Embedding", "Clustering", "k"),
  facetx = NULL,
  facety = NULL,
  limits = NULL
)
}
\arguments{
\item{x}{}

\item{target}{}

\item{x_axis_var}{}

\item{color_var}{}

\item{palette}{}

\item{by}{}

\item{facetx}{}

\item{facety}{}

\item{limits}{}
}
\value{

}
\description{
To help manage the scale differences, outliers (log p < Q1 - 1.5*IQR) are omitted.
}
