% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.cops.R
\name{process.cops}
\alias{process.cops}
\title{This the main function of the Cops package and is
called by \code{\link{cops.go}}}
\usage{
process.cops(dirdat, ASCII = FALSE, CLEAN.FILES = FALSE)
}
\description{
This the main function of the Cops package and is
called by \code{\link{cops.go}}
}
\author{
Bernard Gentili and Simon Belanger
}
