% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_metrics.R
\name{stability_eval}
\alias{stability_eval}
\title{Clustering stability evaluation}
\usage{
stability_eval(clust, by = c("datname", "drname", "run", "k", "m"), ...)
}
\arguments{
\item{clust}{clustering \code{data.frame} such as returned by \code{\link{cv_clusteval}}}

\item{by}{vector of column names to keep}

\item{...}{extra arguments are ignored}
}
\value{
Returns a \code{data.frame} where each row corresponds to clustering stability
        with respect to kept column variables
}
\description{
Performs stability analysis on cross-validated clusterings using \code{\link{jdist}}.
}
\details{
Default settings work with \code{\link{cv_clusteval}} output 'clusters'.
}
