% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{subsampling}
\alias{subsampling}
\title{Subsampling permutations of clustering dataset}
\usage{
subsampling(..., subsampling_strategy = "cv")
}
\arguments{
\item{...}{arguments passed to \code{\link[COPS]{cv_fold}} or \code{\link[COPS]{bootstrap}}}

\item{subsampling_strategy}{either \code{"cv"} or \code{"bootstrap"}}
}
\value{
list of data.frames with added columns "fold", "run" and "cv_index" as well as 
        duplicated rows of the original data corresponding to different folds.
}
\description{
Creates subsamples via cross-validation folds or bootstrapping.
}
