% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_analysis.R
\name{survival_evaluation}
\alias{survival_evaluation}
\title{Likelihood ratio analysis of Cox PH survival models for clustering results}
\usage{
survival_evaluation(
  event_data,
  clusters,
  survival_time_col = "time",
  survival_event_col = "event",
  survival_covariate_names = c("age", "stage"),
  row_id = "ID",
  by = c("run", "fold", "datname", "drname", "k", "m"),
  parallel = 1,
  ...
)
}
\arguments{
\item{event_data}{A data.frame that contains survival times, event data and covariates.}

\item{clusters}{A data.frame or data.table containing clustering information.}

\item{survival_time_col}{Name of the column in \code{event_data} that contains survival time.}

\item{survival_event_col}{Name of the column in \code{event_data} that contains event indicators.}

\item{survival_covariate_names}{Nme of covariate columns.}

\item{row_id}{Name of column in \code{event_data} that matches sample IDs in \code{clusters}.}

\item{by}{Vector of column names that identify a single clustering result in \code{clusters}.}

\item{...}{Extra arguments are ignored.}
}
\value{

}
\description{
Likelihood ratio analysis of Cox PH survival models for clustering results
}
