% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_analysis.R
\name{survival_preprocess}
\alias{survival_preprocess}
\title{Pre-process event data for survival analysis}
\usage{
survival_preprocess(
  event_data,
  event_time_name = "days_to_death",
  follow_up_time_name = "days_to_last_followup",
  event_field_name = "vital_status",
  event_name = "Dead",
  event_time_cutoff = Inf
)
}
\arguments{
\item{event_data}{}

\item{event_time_name}{}

\item{follow_up_time_name}{}

\item{event_field_name}{}

\item{event_name}{}

\item{event_time_cutoff}{}
}
\value{

}
\description{
Pre-process event data for survival analysis
}
