% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{triple_viz}
\alias{triple_viz}
\alias{pca_viz}
\alias{umap_viz}
\alias{tsne_viz}
\title{Data visualization using PCA, t-SNE and UMAP}
\usage{
triple_viz(
  data,
  category,
  category_label,
  tsne_perplexity = 45,
  umap_neighbors = 20,
  tsne = FALSE,
  pre_manifold_pca = TRUE,
  color_scale = scale_color_brewer(palette = "Dark2")
)

pca_viz(
  data,
  category,
  category_label,
  color_scale = scale_color_brewer(palette = "Dark2")
)

umap_viz(
  data,
  category,
  category_label,
  umap_neighbors = 20,
  pre_manifold_pca = TRUE,
  max_pcs = 50,
  color_scale = scale_color_brewer(palette = "Dark2")
)

tsne_viz(
  data,
  category,
  category_label,
  tsne_perplexity = 45,
  pre_manifold_pca = TRUE,
  color_scale = scale_color_brewer(palette = "Dark2")
)
}
\arguments{
\item{data}{\code{matrix} with samples on rows}

\item{category}{\code{factor} for coloring}

\item{category_label}{name of color legend}

\item{tsne_perplexity}{t-SNE perplexity parameter}

\item{umap_neighbors}{UMAP neighbours parameter}

\item{tsne}{whether to use t-SNE}

\item{pre_manifold_pca}{whether to apply PCA before manifold learning 
(recommended for high-dimensional data)}

\item{color_scale}{color scale used for \code{category}}

\item{max_pcs}{maximum number of PCs (limited by data) to extract for UMAP}
}
\value{
list of plots

\code{ggplot} object

\code{ggplot} object

\code{ggplot} object
}
\description{
Generates scatter plots from 2D embeddings.
}
\section{Functions}{
\itemize{
\item \code{pca_viz()}: Data visualization using PCA

\item \code{umap_viz()}: Data visualization using UMAP

\item \code{tsne_viz()}: Data visualization using t-SNE

}}
