% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{triple_viz}
\alias{triple_viz}
\title{Data visualization using PCA, t-SNE and UMAP}
\usage{
triple_viz(
  data,
  category,
  category_label,
  tsne_perplexity = 45,
  umap_neighbors = 20,
  tsne = FALSE
)
}
\arguments{
\item{data}{}

\item{category}{}

\item{category_label}{}

\item{tsne_perplexity}{}

\item{umap_neighbors}{}
}
\description{
Data visualization using PCA, t-SNE and UMAP
}
